% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{ccsDxICD9}
\alias{ccsDxICD9}
\title{Clinical Classifications Software (CCS)  for ICD-9-CM}
\format{A data frame with 15072 rows and 12 variables}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Single_Level_CCS_2015.zip}

\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Multi_Level_CCS_2015.zip}
}
\usage{
data(ccsDxICD9)
}
\description{
A dataframe for ICD-9-CM CCS mapping, containing ICD description, CCS single-level diagnosis classification, and CCS multiple level
}
\details{
\itemize{
  \item ICD. ICD-9-CM
  \item CCS_CATEGORY. single ccs category for ICD-9
  \item ICD_DESCRIPTION. ICD-9 description
  \item CCS_CATEGORY_DESCRIPTION. single ccs description for ICD-9
  \item CCS_LVL_1. multiple ccs level one category for ICD-9
  \item CCS_LVL_1_LABEL. multiple ccs level one description for ICD-9
  \item CCS_LVL_2. multiple ccs level two category for ICD-9
  \item CCS_LVL_2_LABEL. multiple ccs level two description for ICD-9
  \item CCS_LVL_3. multiple ccs level three category for ICD-9
  \item CCS_LVL_3_LABEL. multiple ccs level three description for ICD-9
  \item CCS_LVL_4. multiple ccs level four category for ICD-9
  \item CCS_LVL_4_LABEL. multiple ccs level four description for ICD-9
}
}
\keyword{dataframe}
