% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{ccsPrICD10}
\alias{ccsPrICD10}
\title{Clinical Classifications Software (CCS)  for ICD-10-PCS}
\format{A data frame with 79758 rows and 8 variables}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs10/ccs_pr_icd10pcs_2019_1.zip}
}
\usage{
data(ccsPrICD10)
}
\description{
A datatable for ICD-10-PCS CCS mapping, containing ICD description, CCS single-level procedure classification, and CCS multiple level, which can be used for aggregate statistical reporting of a variety of types (Updated for codes valid through FY 2019).
}
\details{
\itemize{
  \item ICD. ICD-10-PCS
  \item CCS_CATEGORY. single ccs category for ICD-10
  \item ICD_DESCRIPTION. ICD-10 description
  \item CCS_CATEGORY_DESCRIPTION. single ccs description for ICD-10
  \item CCS_LVL_1. multiple ccs level one category for ICD-10
  \item CCS_LVL_1_LABEL. multiple ccs level one description for ICD-10
  \item CCS_LVL_2. multiple ccs level two category for ICD-10
  \item CCS_LVL_2_LABEL. multiple ccs level two description for ICD-10
}
}
\seealso{
\code{\link{ccsPrICD9}}
}
\keyword{datatable}
