% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{ccsPrICD9}
\alias{ccsPrICD9}
\title{Clinical Classifications Software (CCS)  for ICD-9-PCS}
\format{A data table with 3948 rows and 10 variables}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Single_Level_CCS_2015.zip}

\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Multi_Level_CCS_2015.zip}
}
\usage{
data(ccsPrICD9)
}
\description{
A datatable for ICD-9-PCS CCS mapping, containing ICD description, CCS single-level procedure classification, and CCS multiple level, which can be used for aggregate statistical reporting of a variety of types. (Updated for codes valid through FY 2015.)
}
\details{
\itemize{
  \item ICD. ICD-9-PCS
  \item CCS_CATEGORY. single ccs category for ICD-9
  \item ICD_DESCRIPTION. ICD-9 description
  \item CCS_CATEGORY_DESCRIPTION. single ccs description for ICD-9
  \item CCS_LVL_1. multiple ccs level one category for ICD-9
  \item CCS_LVL_1_LABEL. multiple ccs level one description for ICD-9
  \item CCS_LVL_2. multiple ccs level two category for ICD-9
  \item CCS_LVL_2_LABEL. multiple ccs level two description for ICD-9
  \item CCS_LVL_3. multiple ccs level three category for ICD-9
  \item CCS_LVL_3_LABEL. multiple ccs level three description for ICD-9
}
}
\seealso{
\code{\link{ccsPrICD10}}
}
\keyword{datatable}
