% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R
\name{common_PrArg}
\alias{common_PrArg}
\title{Common argument of procedure functions}
\arguments{
\item{PrDataFile}{A file of clinical procedure data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A MemberID column of PrDataFile}

\item{icdColName}{A ICD column of PrDataFile}

\item{dateColName}{A Date column of PrDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{isDescription}{Category or description of standard classification methods for ICD procedure codes. By default, it set to \code{True} (standard description).}
}
\description{
Common argument of procedure functions
}
\details{
Common argument of procedure functions
}
