% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIcdDecimaltoShort.R
\name{convertIcdDecimaltoShort}
\alias{convertIcdDecimaltoShort}
\title{Convert ICD Codes From Decimal To Short Forms}
\usage{
convertIcdDecimaltoShort(icdList, icdVer)
}
\arguments{
\item{icdList}{ICD codes}

\item{icdVer}{ICD Version: ICD-9 or ICD-10}
}
\description{
Convert codes between short and decimal forms
}
