% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIcdDxDecimaltoShort.R
\name{convertIcdDxDecimaltoShort}
\alias{convertIcdDxDecimaltoShort}
\title{Convert ICD Codes From Decimal To Short Forms}
\usage{
convertIcdDxDecimaltoShort(icdList)
}
\arguments{
\item{icdList}{ICD codes}
}
\description{
Convert codes between short and decimal forms
}
