% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIcdDxShortToDecimal.R
\name{convertIcdDxShortToDecimal}
\alias{convertIcdDxShortToDecimal}
\title{Convert ICD Codes From Short To Decimal Forms}
\usage{
convertIcdDxShortToDecimal(icdList)
}
\arguments{
\item{icdList}{ICD codes}
}
\description{
Convert codes between short and decimal forms
}
