% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIcdPrDecimaltoShort.R
\name{convertIcdPrDecimaltoShort}
\alias{convertIcdPrDecimaltoShort}
\title{Convert ICD-9-PCS Codes From Decimal To Short Forms}
\usage{
convertIcdPrDecimaltoShort(icdList)
}
\arguments{
\item{icdList}{only ICD-9-PCS codes have decimal and short forms}
}
\description{
Return codes with short forms
}
