% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIcdShortToDecimal.R
\name{convertIcdShortToDecimal}
\alias{convertIcdShortToDecimal}
\title{Convert ICD Codes From Short To Decimal Forms}
\usage{
convertIcdShortToDecimal(icdList, icdVer)
}
\arguments{
\item{icdList}{ICD codes}

\item{icdVer}{ICD Version: ICD-9 or ICD-10}
}
\description{
Convert codes between short and decimal forms
}
