% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_emr.R
\docType{data}
\name{Dx_dataset_classification}
\alias{Dx_dataset_classification}
\alias{ccsDxICD9}
\alias{ccsDxICD10}
\alias{phecode_icd9_2}
\alias{icd9_ahrq}
\alias{icd10_ahrq}
\alias{icd9_charlson}
\alias{icd10_charlson}
\alias{icd9_elix}
\alias{icd10_elix}
\title{Datasets for code classification}
\format{
These datatable with ICD diagnostic codes and the specific Standardization categories.

An object of class \code{data.table} (inherits from \code{data.frame}) with 72446 rows and 8 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 15558 rows and 9 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 1487 rows and 3 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 3217 rows and 3 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 1241 rows and 3 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 2804 rows and 3 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 1474 rows and 3 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 3493 rows and 3 columns.
}
\usage{
ccsDxICD9

ccsDxICD10

phecode_icd9_2

icd9_ahrq

icd10_ahrq

icd9_charlson

icd10_charlson

icd9_elix

icd10_elix
}
\description{
These datasets are used for diagnostic code transformation.
}
\details{
These datasets are used for \code{IcdDxToCCS}, \code{IcdDxToCCSLvl}, \code{IcdDxToPheWAS}, and \code{IcdDxToComorbid}
}
\keyword{datasets}
