% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_emr.R
\docType{data}
\name{Dx_dataset_standardization}
\alias{Dx_dataset_standardization}
\alias{ICD9DxwithTwoFormat}
\alias{ICD10DxwithTwoFormat}
\title{Datasets for code standardization}
\format{These datatable with ICD9 and ICD10 diagnosis codes.}
\usage{
ICD9DxwithTwoFormat

ICD10DxwithTwoFormat
}
\description{
These datasets are used for diagnostic code standardization
}
\details{
These datasets are used for \code{icdDxShortToDecimal}, and \code{icdDxDecimalToShort}. Transform the diagnosis codes into uniform format before the integration process.
}
\keyword{datasets}
