% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxToComorbid.R, R/fun_emr.R
\name{IcdDxToComorbid}
\alias{IcdDxToComorbid}
\alias{DxComorbid}
\title{Code classification for Comorbidity}
\source{
AHRQ

ICD-9-CM Elixhauser (2012-2015)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp#references}

ICD-10-CM Elixhauser (2019)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity_icd10.jsp}

Charlson

ICD-9-CM Charlson (2006)

\url{http://mchp-appserv.cpe.umanitoba.ca/Upload/SAS/ICD9_E_Charlson.sas.txt}

ICD-10-CM Charlson (2006)

\url{http://mchp-appserv.cpe.umanitoba.ca/Upload/SAS/ICD10_Charlson.sas.txt}

Elixhauser

ICD-9-CM Elixhauser (2012-2015)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp#references}

ICD-10-CM Elixhauser (2019)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity_icd10.jsp}
}
\usage{
IcdDxToComorbid(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, comorbidMethod, isDescription = FALSE)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of DxDataFile.}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{comorbidMethod}{Three comorbidity method: AHRQ, Charlson and Elixhauser Comorbidity. Change it to any of the other possible variables (\code{ahrq},\code{charlson}, and \code{elix}).}

\item{isDescription}{Category or description of standard classification methods for ICD diagnostic codes. By default it is set to \code{FALSE} ( Comorbidity categories).}
}
\description{
The comorbidities classification (AHRQ, Charlson ,and Elixhauser Comorbidity) for ICD diagnostic codes is a diagnostic categorization scheme that can employ in many types of projects analyzing data on diagnoses.
}
\details{
Collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than are individual ICD diagnostic codes.
}
\examples{
# sample file for example

head(sampleDxFile)

# Group diagnostic codes into charlson comorbidity categories

IcdDxToComorbid(sampleDxFile, ID, ICD, Date, "2015-10-01", charlson)
}
