% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxToCustom.R, R/IcdDxToCustomGrep.R,
%   R/fun_emr.R
\name{IcdDxToCustom}
\alias{IcdDxToCustom}
\alias{IcdDxToCustomGrep}
\alias{DxCustom}
\title{Code classification for customized group}
\usage{
IcdDxToCustom(DxDataFile, idColName, icdColName, dateColName,
  CustomGroupingTable)

IcdDxToCustomGrep(DxDataFile, idColName, icdColName, dateColName,
  CustomGroupingTable)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of DxDataFile.}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{CustomGroupingTable}{Define the grouping categories. `IcdDxToCustom` needs a dataset with two columns: "Group", "ICD"; "Group" defines one or more disease categories. "ICD" defines a list of disease-related ICD codes. `IcdDxToCustomGrep`needs a dataset with two columns: "Group", "grepIcd"; "Group" defines one or more disease categories. "grepICD" defines disease-related ICD code character strings containing a regular expression.}
}
\value{
Two new \code{data.table}s. 1) \code{groupedDT}: Based on \code{DxDataFile} with two new columns for uniform format diagnostic codes and classified standard categories. 2) \code{summarised_groupedDT}: Summarized the dataset  \code{groupedDT} and sorted by memberID.
}
\description{
Researches can define the grouping categories and therefore have more flexible for grouping ICD diagnostic codes.
}
\details{
There are two functions for customized defined grouping method, the customized category grouping is based on precise (`IcdDxToCustom`) and fuzzy (`IcdDxToCustomGrep`) grouping method, respectively.
}
\examples{
# sample file for example

head(sampleDxFile)

# Group diagnostic codes into "Chronic kidney disease" with precise grouping method

groupingTable <- data.frame(Group = rep("Chronic kidney disease",6),
                            ICD = c("N181","5853","5854","5855","5856","5859"),
                            stringsAsFactors = FALSE)

IcdDxToCustom(sampleDxFile, ID, ICD, Date, CustomGroupingTable = groupingTable)

# Group diagnostic codes into "Chronic kidney disease" with fuzzy grouping method

grepTable <- data.frame(Group = "Chronic kidney disease",
                        grepIcd = "^585|^N18",
                        stringsAsFactors = FALSE)

IcdDxToCustomGrep(sampleDxFile,ID, ICD, Date,CustomGroupingTable = grepTable)

}
\seealso{
Other code classification functions: \code{\link{DxPheWAS}}, \code{\link{DxComorbid}}, \code{\link{DxCCS}}
}
