% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{dxICD10}
\alias{dxICD10}
\title{ICD-10-CM}
\format{A data frame with 71931 rows and 2 variables}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD10/2019-ICD-10-CM.html}
}
\usage{
data(dxICD10)
}
\description{
A dataframe for ICD-10-CM, containing ICD-10-CM code and ICD description(Updated for codes valid through FY 2019.)
}
\details{
\itemize{
  \item ICD. ICD-10-CM
  \item ICD_DESCRIPTION. ICD-10-CM description
}
}
\seealso{
\code{\link{dxICD9}}
}
\keyword{dataframe}
