% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{dxICD9}
\alias{dxICD9}
\title{ICD-9-CM}
\format{A data frame with 14567 rows and 2 variables}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}
}
\usage{
data(dxICD9)
}
\description{
A datatable for ICD-9-CM, containing ICD-9-CM code and ICD description (Updated for codes valid through FY 2011).
}
\details{
\itemize{
  \item ICD. ICD-9-CM
  \item ICD_DESCRIPTION. ICD-9-CM description
}
}
\seealso{
\code{\link{dxICD10}}
}
\keyword{datatable}
