% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxToPheWAS.R, R/fun_emr.R
\name{IcdDxToPheWAS}
\alias{IcdDxToPheWAS}
\alias{DxPheWAS}
\title{Code classification for PheWAS}
\usage{
IcdDxToPheWAS(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, isDescription = TRUE)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of DxDataFile.}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{isDescription}{Category or description of standard classification methods for ICD diagnostic codes. By default, it set to \code{True} (standard description).}
}
\value{
Three new \code{data.table}s. 1) \code{groupedDT}: Based on \code{DxDataFile} with two new columns for uniform format diagnostic codes and classified standard categories. 2) \code{summarised_groupedDT}: Summarized the dataset  \code{groupedDT} and sorted by memberID. 3) \code{Error}: Potential error codes from \code{\link{DxUniform}}.
}
\description{
The PheWAS classification for ICD-9-CM codes is a diagnostic categorization scheme that can employ in many types of projects analyzing data on diagnoses.
}
\details{
Collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than are individual ICD diagnostic codes.
}
\examples{
# sample file for example

head(sampleDxFile)

# Group diagnostic codes into PheWAS

IcdDxToPheWAS(sampleDxFile, ID, ICD, Date, "2015-10-01", FALSE)
}
\seealso{
Other code classification functions: \code{\link{DxCustom}}, \code{\link{DxComorbid}}, \code{\link{DxCCS}}
}
