% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/icdDxDecimalToShort.R,
%   R/icdDxShortToDecimal.R
\name{dxUniform}
\alias{dxUniform}
\alias{icdDxDecimalToShort}
\alias{icdDxShortToDecimal}
\title{Code format transformation}
\usage{
icdDxDecimalToShort(dxDataFile, icdColName, dateColName,
  icdVerColName = NULL, icd10usingDate = NULL)

icdDxShortToDecimal(dxDataFile, icdColName, dateColName,
  icdVerColName = NULL, icd10usingDate = NULL)
}
\arguments{
\item{dxDataFile}{A file of clinical diagnostic data with at least 2 columns: "ICD" and "Date"}

\item{icdColName}{A column for ICD of dxDataFile}

\item{dateColName}{A column for Date of dxDataFile}

\item{icdVerColName}{A column for ICD Version used of dxDataFile. If this argument is provided, then icd10usingDate is not needed.}

\item{icd10usingDate}{The date that ICD-10 was started using}
}
\value{
Two new \code{data.table}s. 1) \code{ICD}: Uniform format diagnostic codes. 2) \code{Error}: Potential error codes.
}
\description{
Convert the ICD diagnostic codes to a uniform format
}
\details{
`icdDxShortToDecimal` is used for grouping diagnostic code to PheWAS classification (icdDxToPheWAS). `IcdDxDecimalToShort` can be used for grouping to the others classification (icdDxToCCS, icdDxToCCSLvl, and icdDxToComorbid). The transformative function can not only convert the ICD to uniform format code but also check potential coding error of the ICD codes’ format or version.
}
\examples{
# sample file for example

head(sampleDxFile)

# convert the diagnostic codes to the decimal format

icdDxShortToDecimal(dxDataFile = sampleDxFile, icdColName = ICD, dateColName = Date, icdVerColName = NULL, icd10usingDate = "2015/10/01")

# convert the diagnostic codes to the short format

icdDxDecimalToShort(dxDataFile = sampleDxFile, icdColName = ICD, dateColName = Date, icdVerColName = NULL, icd10usingDate = "2015/10/01")
}
