% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/getEligiblePeriod.R
\name{eligiblePeriod}
\alias{eligiblePeriod}
\alias{getEligiblePeriod}
\title{Data integration for patient record period}
\usage{
getEligiblePeriod(dxDataFile, idColName, dateColName)
}
\arguments{
\item{dxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of dxDataFile.}

\item{dateColName}{A column for Date of dxDataFile}
}
\value{
A new \code{data.table} based on \code{dxDataFile} with the earliest and latest admission date for each patient.
}
\description{
Function for researchers used for finding the first and last clinical event for a given patient as index date.
}
\details{
The function queries the earliest and latest admission date for each patient.
}
\examples{
# sample file for example

head(sampleDxFile)

# Earliest and latest admission date for each patient

record <- getEligiblePeriod(sampleDxFile, ID, Date)
head(record)
}
\seealso{
Other data integration functions: \code{\link{selectCases}}, \code{\link{dataSplit}}, \code{\link{era}}
}
