% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConditionEra.R
\name{getConditionEra}
\alias{getConditionEra}
\title{Get the condition era}
\usage{
getConditionEra(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, gapDate = 30, conditionSelect = CCS,
  isDescription = TRUE, CustomGroupingTable)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{Icd 10 using date}

\item{gapDate}{Length of condition era, default is 30 days}

\item{conditionSelect}{Four Stratified method can be chosen: CCS (or CCS levels), phecode (only for icd-9), comorbidities, grepICD or customICD , type `ccs`,`ccslvl1`,`ccslvl2`,`ccslvl3`,`ccslvl4`, `phecode`, `ahrq`, `charlson`, `elix` `customGrepIcdGroup`, or `customIcdGroup`, default is CCS}

\item{isDescription}{CCS/Phecode categories or description for ICD-CM codes, default is True}

\item{CustomGroupingTable}{Table is for groupedICDMethod:`grepICD` and `customGroup`}
}
\description{
A Condition Era is defined as a span of time when the member is assumed to have a given condition.
Condition Eras are periods of Condition Occurrence.
Combining individual Condition Occurrences into a single Condition Era based on ICD code in clinical diagnostic data.
Condition Eras are built with a Persistence Window,deafault is 30 days, meaning, if no occurence of the same member id happens within 30 days of any one occurrence, it will be considered the end date of the last condition occurrence.
}
\details{
return DxDataFile with new column, condition era.
}
\examples{
head(sampleDxFile)
getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, ccs, FALSE)
getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, ICD)
getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, phecode, FALSE)
getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, elix)
grepTable <- data.frame(group = "Cardiac dysrhythmias",
                        grepIcd = "^427|^I48",
                        stringsAsFactors = FALSE)
getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01", 30, customGrepIcdGroup,
                CustomGroupingTable = grepTable)

}
