% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/getEligiblePeriod.R
\name{getEligiblePeriod}
\alias{getEligiblePeriod}
\title{Data integration for patient record period}
\usage{
getEligiblePeriod(dxDataFile, idColName, dateColName)
}
\arguments{
\item{dxDataFile}{A data frame object of clinical diagnostic data with at least 3 columns: ID, ICD, and Date. As for date column, the data format should be YYYY/MM/DD or YYYY-MM-DD.}

\item{idColName}{Column name of ID column in dxDataFile. Data type of this argumant should be string without quotation marks.}

\item{dateColName}{Column name of date column in dxDataFile (with date in YYYY/MM/DD or YYYY-MM-DD format). Data type of this argumant should be string without quotation marks.}
}
\value{
A new \code{data.table} based on \code{dxDataFile} with the earliest and latest admission date for each patient.
}
\description{
\code{getEligiblePeriod} is a function to find the first and last clinical event for a given patient. The result can be index date in the folloing process such as \code{\link{splitDataByDate}}.
}
\details{
The function queries the earliest and latest admission date for each patient.
}
\examples{
# sample file for example

head(sampleDxFile)

# Earliest and latest admission date for each patient

record <- getEligiblePeriod(sampleDxFile, ID, Date)
head(record)
}
\seealso{
Other data integration functions: \code{\link{selectCases}}, \code{\link{splitDataByDate}}, \code{\link{getConditionEra}}
}
