% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupIcdBasedOnCCS.R
\name{groupIcdBasedOnCCS}
\alias{groupIcdBasedOnCCS}
\title{Get the Clinical Classifications Software (CCS) single and multiple category/ description}
\usage{
groupIcdBasedOnCCS(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, isCCSCategoryDescription = TRUE)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{isCCSCategoryDescription}{Clinical Classifications Software (CCS) single level categories/ description for ICD-9 or ICD-10, default is True}
}
\description{
This can be used to select the Clinical Classifications Software (CCS) single and multiple category/ description
based on ICD code in clinical diagnostic data,
return Clinical Classifications Software (CCS) single and multiple category/ description based on ICD-9 and ICD-10
}
\examples{
DxDataFile <- data.frame(ID=c("A","A","A"),
                         ICD=c("6929","V433","I350"),
                         Date=as.Date(c("2013-03-31","2013-01-29","2016-03-10")),
                         stringsAsFactors = F)
groupIcdBasedOnCCS (DxDataFile, ID, ICD, Date, "2016-01-01", T)
}
