% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupIcdBasedOnComorbid.R
\name{groupIcdBasedOnComorbid}
\alias{groupIcdBasedOnComorbid}
\title{Grouping comorbid method comorbidities (AHRQ, Charlson and Elixhauser Comorbidity) infers whether to use ICD-9 or ICD-10 codes}
\usage{
groupIcdBasedOnComorbid(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, comorbidMethod, NumericOrBinary = B, groupByDate = T)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{Icd 10 using date}

\item{comorbidMethod}{Three comorbidity method: AHRQ, Charlson and Elixhauser Comorbidity, type `ahrq`,`charlson`, or`elix`}

\item{NumericOrBinary}{Member have one (or more) diagnostic comorbidities, type `N` or `B`, default is `B` (Binary)}

\item{groupByDate}{Default is True.}
}
\description{
Get comorbidities using the comorbidity methods based on ICD code in clinical diagnostic data.
}
\details{
return comorbidity meseaures based on ICD diagnosis codes
}
\examples{
DxDataFile <- data.frame(ID=c("A","A","B","B"),
                         ICD=c("40201","V433","I350","I350"),
                         Date=as.Date(c("2013-03-31","2013-01-29","2016-03-10","2016-03-10")),
                         stringsAsFactors = F)

groupIcdBasedOnComorbid(DxDataFile, ID, ICD, Date, "2016-01-01", ahrq, N, T)
groupIcdBasedOnComorbid(DxDataFile, ID, ICD, Date, "2016-01-01", charlson, B, T)
groupIcdBasedOnComorbid(DxDataFile, ID, ICD, Date, "2016-01-01", elix, N, T)

}
