% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas.R
\name{groupICDBasedOnPhecode}
\alias{groupICDBasedOnPhecode}
\title{Get the PheWAS Phecode/ description}
\usage{
groupICDBasedOnPhecode(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, isPhecodeDescription = TRUE)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{icd 10 using date}

\item{isPhecodeDescription}{PheWAS Phecode/ description for icd9, default is True}
}
\description{
This can be used to select the PheWAS Phecode/ description
based on ICD code in clinical diagnostic data,
return CCS single and multiple category/ description based on ICD
}
\examples{
groupICDBasedOnPhecode(DxDataFile, ID, ICD, Date, "2016-01-01", F)

}
