% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupIcdPrToCCSLvl.R
\name{groupIcdPrToCCSLvl}
\alias{groupIcdPrToCCSLvl}
\title{Get the Multi-level Clinical Classifications Software (CCS) categories and description for ICD-9 and ICD-10 codes on procedures.}
\usage{
groupIcdPrToCCSLvl(PrDataFile, idColName, icdColName, dateColName,
  icd10usingDate, CCSLevel = 1, CCSLvlLabel = TRUE)
}
\arguments{
\item{PrDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", "Date"}

\item{idColName}{A column for MemberID of PrDataFile}

\item{icdColName}{A column for ICD of PrDataFile}

\item{dateColName}{A column for Date of PrDataFile}

\item{icd10usingDate}{icd 10 using date}

\item{CCSLevel}{Clinical Classifications Software (CCS) multiple level}

\item{CCSLvlLabel}{Clinical Classifications Software (CCS) multiple level categories/description for icd9/10, default is True}
}
\description{
Multi-leve Clinical Classifications Software (CCS) for ICD-9 and ICD-10 procedure codes in clinical procedures data is a procedure categorization scheme. Four levels exist in the multi-level procedure CCS for ICD-9-CM codes, and two levels exist in the multi-level procedure CCS for ICD-10-CM codes
}
\details{
return Multi-leve Clinical Classifications Software (CCS) categories or description based on ICD-9 and ICD-10 codes
}
\examples{
PrDataFile <- data.frame(ID=c("A","A","A","B"),
                         ICD=c("0101","8838","00870ZZ","00920ZZ"),
                         Date=as.Date(c("2013-03-31","2013-01-29","2016-03-10","2016-03-10")),
                         stringsAsFactors = FALSE)
groupIcdPrToCCSLvl(PrDataFile, ID, ICD, Date, "2016-01-01",2, TRUE)

}
