% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupIcdToCustomGrep.R
\name{groupIcdToCustomGrep}
\alias{groupIcdToCustomGrep}
\title{Get the categories of ICD-9 and ICD-10 codes on diagnoses, the grouping rules are based on your standards.}
\usage{
groupIcdToCustomGrep(icdFile, groupingTable)
}
\arguments{
\item{icdFile}{ICD-9 and ICD-10 codes}

\item{groupingTable}{Grouping rules of clustering the ICD is based on yourself! There are two column in the dataframe: Group, GrepICD#'}
}
\description{
This can be used to select the first diagnosis record
based on ICD code (grepl language) in DxDataFile,
return first diagnosis record based on factIcd Data
}
\examples{
icdFile <- data.frame(ICD = c("I95.0", "I952", "I110", "01091"), stringsAsFactors = FALSE)
groupingTable <- data.frame(Group = c("Hypotension","Hypotension", "Hypertension", "Hypertension"),
                            ICD = c("I95.0","I951","I110","I11.9"),
                            stringsAsFactors = FALSE)
groupIcdToCustomGrep(icdFile, groupingTable)

}
