% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupIcdToPhecode.R
\name{groupIcdToPhecode}
\alias{groupIcdToPhecode}
\title{Get the Phecode or description of ICD-9 diagnosis codes}
\source{
\url{https://phewascatalog.org/phecodes}
}
\usage{
groupIcdToPhecode(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, isPhecodeDescription = TRUE)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{icd 10 using date}

\item{isPhecodeDescription}{Phecode/ description for icd9, default is True}
}
\description{
This can be used to group Phecode or description based on ICD-9 codes in clinical diagnostic data.
}
\details{
return Phecode or description based on ICD-9-CM codes
}
\examples{
DxDataFile <- data.frame(ID = c("A", "A", "A"),
                         ICD = c("6929", "V433", "I350"),
                         Date = as.Date(c("2013-03-31", "2013-01-29", "2016-03-10")),
                         stringsAsFactors = FALSE)
groupIcdToPhecode(DxDataFile, ID, ICD, Date, "2016-01-01", FALSE)

}
