% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{icd10_ahrq}
\alias{icd10_ahrq}
\title{AHRQ Comorbidity for ICD-10}
\format{A data table with 3217 rows and 3 variables}
\usage{
data(icd10_ahrq)
}
\description{
A datatable for ICD-10 AHRQ comorbidity mapping
}
\details{
\itemize{
  \item ICD. ICD-10
  \item Comorbidity. comorbidity measures for ICD-10 (CHF, Valvular, PHTN, PVD, HTN, HTNcx, Paralysis, NeuroOther, Pulmonary, DM, DMcx, Hypothyroid, Renal, Liver, PUD, HIV, Lymphoma, Mets, Tumor, Rheumatic, Coagulopathy, Obesity, WeightLoss, FluidsLytes, BloodLoss, Anemia, Alcohol, Drugs, Psychoses, Depression)
  \item Value. if ICD code has comorbidity measures, value is 1
}
}
\seealso{
\code{\link{icd9_ahrq}} \code{\link{icd9_charlson}} \code{\link{icd10_charlson}} \code{\link{icd9_elix}} \code{\link{icd10_elix}}
}
\keyword{datatable}
