% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{icd10_charlson}
\alias{icd10_charlson}
\title{Charlson Comorbidity for ICD-10}
\format{A data frame with 3029 rows and 3 variables}
\usage{
data(icd10_charlson)
}
\description{
A dataframe for ICD-10 Charlson comorbidity mapping
}
\details{
\itemize{
  \item ICD. ICD-10
  \item Comorbidity. comorbidity measures for ICD-10 (MI, CHF, PVD, Stroke, Dementia, Pulmonary, Rheumatic, PUD, LiverMild, DM, DMcx, Paralysis, Renal, Cancer, LiverSevere, Mets, HIV)
  \item Value. if ICD code has comorbidity measures, value is 1
}
}
\seealso{
\code{\link{icd9_ahrq}} \code{\link{icd10_ahrq}} \code{\link{icd9_charlson}} \code{\link{icd9_elix}} \code{\link{icd10_elix}}
}
\keyword{dataframe}
