% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{ICD10DxwithTwoFormat}
\alias{ICD10DxwithTwoFormat}
\title{Two format of ICD-10-CM (2019): Short and Decimal}
\format{A data table with 71931 rows and 2 variables}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD10/2019-ICD-10-CM.html}
}
\usage{
data(ICD10DxwithTwoFormat)
}
\description{
A dataframe for ICD-10 format conversion
}
\details{
\itemize{
  \item Short ICD-10 Short format
  \item Decimal ICD-10 Decimal format
}
}
\seealso{
\code{\link{ICD9DxwithTwoFormat}}
}
\keyword{datatable}
