% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{icd10withTwoFormat}
\alias{icd10withTwoFormat}
\title{Two format of ICD-10-CM: Short and Decimal}
\format{A data frame with 72167 rows and 2 variables}
\usage{
data(icd10withTwoFormat)
}
\description{
A dataframe for ICD-10 format conversion
}
\details{
\itemize{
  \item Short ICD-10 Short format
  \item Decimal ICD-10 Decimal format
}
}
\keyword{dataframe}
