% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{icd9_ahrq}
\alias{icd9_ahrq}
\title{AHRQ Comorbidity for ICD-9}
\format{A data frame with 14696 rows and 3 variables}
\usage{
data(icd9_ahrq)
}
\description{
A dataframe for ICD-9 AHRQ comorbidity mapping
}
\details{
\itemize{
  \item ICD. ICD-9
  \item Comorbidity. comorbidity measures for ICD-9 (CHF, Valvular, PHTN, PVD, HTN, HTNcx, Paralysis, NeuroOther, Pulmonary, DM, DMcx, Hypothyroid, Renal, Liver, PUD, HIV, Lymphoma, Mets, Tumor, Rheumatic, Coagulopathy, Obesity, WeightLoss, FluidsLytes, BloodLoss, Anemia, Alcohol, Drugs, Psychoses, Depression)
  \item Value. if ICD code has comorbidity measures, value is 1
}
}
\keyword{dataframe}
