% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{icd9_elix}
\alias{icd9_elix}
\title{Elixhauser Comorbidity for ICD-9}
\format{A data frame with 13742 rows and 3 variables}
\usage{
data(icd9_elix)
}
\description{
A dataframe for ICD-9 Elixhauser comorbidity mapping
}
\details{
\itemize{
  \item ICD. ICD-9
  \item Comorbidity. comorbidity measures for ICD-9 (CHF, Arrhythmia, Valvular, PHTN, PVD, HTN, HTNcx, Paralysis, NeuroOther, Pulmonary, DM, DMcx, Hypothyroid, Renal, Liver, PUD, HIV, Lymphoma, Mets, Tumor, Rheumatic, Coagulopathy, Obesity, WeightLoss, FluidsLytes, BloodLoss, Anemia, Alcohol, Drugs, Psychoses, Depression)
  \item Value. if ICD code has comorbidity measures, value is 1
}
}
\seealso{
\code{\link{icd9_ahrq}} \code{\link{icd10_ahrq}} \code{\link{icd9_charlson}} \code{\link{icd10_charlson}} \code{\link{icd10_elix}}
}
\keyword{dataframe}
