% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{ICD9DxwithTwoFormat}
\alias{ICD9DxwithTwoFormat}
\title{Two format of ICD-9-CM (2011): Short and Decimal}
\format{A data frame with 14567 rows and 2 variables}
\source{
\url{https://www.findacode.com/search/search.php}

\url{https://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}

\url{https://www.cms.gov/Medicare/Quality-Initiatives-Patient-Assessment-Instruments/HospitalQualityInits/Downloads/HospitalAppendix_F.pdf}
}
\usage{
data(ICD9DxwithTwoFormat)
}
\description{
A dataframe for ICD-9 format conversion
}
\details{
\itemize{
  \item Short ICD-9 Short format
  \item Decimal ICD-9 Decimal format
}
}
\seealso{
\code{\link{ICD10DxwithTwoFormat}}
}
\keyword{dataframe}
