% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd9GenerateDecimalFormat.R
\name{icd9GenerateDecimalFormat}
\alias{icd9GenerateDecimalFormat}
\title{get ICD-9-CM Codes Decimal Forms}
\usage{
icd9GenerateDecimalFormat(icd9)
}
\arguments{
\item{icd9}{ICD-9-CM codes}
}
\description{
Convert codes between short and decimal forms
}
