% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{icd9withTwoFormat}
\alias{icd9withTwoFormat}
\title{Two format of ICD-9-CM: Short and Decimal}
\format{A data frame with 15072 rows and 2 variables}
\usage{
data(icd9withTwoFormat)
}
\description{
A dataframe for ICD-9 format conversion
}
\details{
\itemize{
  \item Short ICD-9 Short format
  \item Decimal ICD-9 Decimal format
}
}
\keyword{dataframe}
