% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICD_10_CM_GenerateDecimalFormat.R
\name{ICD_10_CM_GenerateDecimalFormat}
\alias{ICD_10_CM_GenerateDecimalFormat}
\title{get ICD-10-CM Codes Decimal Forms}
\source{
\url{https://www.findacode.com/search/search.php}
}
\usage{
ICD_10_CM_GenerateDecimalFormat(icd10)
}
\arguments{
\item{icd10}{ICD-10-CM codes}
}
\description{
Convert codes between short and decimal forms.
Format ICD-10-CM Principal and Other Diagnosis Codes can be found at url link.
}
