% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{ICD_10_CM_withTwoFormat}
\alias{ICD_10_CM_withTwoFormat}
\title{Two format of ICD-10-CM: Short and Decimal}
\format{A data frame with 72167 rows and 2 variables}
\source{
\url{https://www.findacode.com/search/search.php}
}
\usage{
data(ICD_10_CM_withTwoFormat)
}
\description{
A dataframe for ICD-10 format conversion
}
\details{
\itemize{
  \item Short ICD-10 Short format
  \item Decimal ICD-10 Decimal format
}
}
\seealso{
\code{\link{ICD_9_CM_withTwoFormat}}
}
\keyword{dataframe}
