% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICD_9_PCS_GenerateDecimalFormat.R
\name{ICD_9_PCS_GenerateDecimalFormat}
\alias{ICD_9_PCS_GenerateDecimalFormat}
\title{get ICD-9-PCS Codes Decimal Forms}
\source{
\url{https://www.cms.gov/Medicare/Quality-Initiatives-Patient-Assessment-Instruments/HospitalQualityInits/Downloads/HospitalAppendix_F.pdf}
}
\usage{
ICD_9_PCS_GenerateDecimalFormat(icd9)
}
\arguments{
\item{icd9}{ICD-9-PCS codes}
}
\description{
Convert codes between short and decimal forms
Format ICD-9-PCS Principal and Other Diagnosis Codes can be found at url link.
}
