% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxDecimaltoShort.R
\name{IcdDxDecimaltoShort}
\alias{IcdDxDecimaltoShort}
\title{Convert ICD Codes From Decimal To Short Forms}
\usage{
IcdDxDecimaltoShort(icdList)
}
\arguments{
\item{icdList}{ICD codes}
}
\description{
Convert codes between short and decimal forms
}
