% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxToCCS.R
\name{IcdDxToCCS}
\alias{IcdDxToCCS}
\title{Get the Clinical Classifications Software (CCS) categories and description for ICD-9 and ICD-10 codes on diagnoses.}
\source{
ICD-9-CM CCS (2012)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Single_Level_CCS_2015.zip}

\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Multi_Level_CCS_2015.zip}

ICD-10-CM CCS (2019)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs10/ccs_dx_icd10cm_2019_1.zip}
}
\usage{
IcdDxToCCS(DxDataFile, idColName, icdColName, dateColName, icd10usingDate,
  isCCSCategoryDescription = T)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{isCCSCategoryDescription}{Clinical Classifications Software (CCS) single level categories/description for ICD-9 or ICD-10, default is True}
}
\description{
Clinical Classifications Software (CCS) for ICD-9 and ICD-10 diagnosis codes in clinical diagnostic data is a diagnosis categorization scheme.
}
\details{
return Clinical Classifications Software (CCS) categories or description based on ICD-9 and ICD-10 codes
}
\examples{

IcdDxToCCS(testDxFile, ID, ICD, Date, "2015-10-01", TRUE)

}
