% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxToComorbid.R
\name{IcdDxToComorbid}
\alias{IcdDxToComorbid}
\title{Grouping comorbid method comorbidities (AHRQ, Charlson and Elixhauser Comorbidity) infers whether to use ICD-9 or ICD-10 codes}
\source{
AHRQ

Charlson

Elixhauser

ICD-9-CM Elixhauser (2012-2015)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp#references}

ICD-10-CM Elixhauser (2019)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity_icd10.jsp}
}
\usage{
IcdDxToComorbid(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, comorbidMethod, NumericOrBinary = B,
  groupByDate = TRUE)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{Icd 10 using date}

\item{comorbidMethod}{Three comorbidity method: AHRQ, Charlson and Elixhauser Comorbidity, type `ahrq`,`charlson`, or`elix`}

\item{NumericOrBinary}{Member have one (or more) diagnostic comorbidities, type `N` or `B`, default is `B` (Binary)}

\item{groupByDate}{Default is True.}
}
\description{
Get comorbidities using the comorbidity methods based on ICD code in clinical diagnostic data.
}
\details{
return comorbidity meseaures based on ICD diagnosis codes
}
\examples{
DxDataFile <- data.frame(ID=c("A","A","B","B"),
                         ICD=c("40201","42577","I350","K289"),
                         Date=as.Date(c("2013-03-31","2013-01-29","2016-03-10","2016-03-10")),
                         stringsAsFactors = FALSE)

IcdDxToComorbid(DxDataFile, ID, ICD, Date, "2016-01-01", ahrq, N, TRUE)
IcdDxToComorbid(DxDataFile, ID, ICD, Date, "2016-01-01", charlson, B, TRUE)
IcdDxToComorbid(DxDataFile, ID, ICD, Date, "2016-01-01", elix, N, TRUE)

}
