% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxToCustomGrep.R
\name{IcdDxToCustomGrep}
\alias{IcdDxToCustomGrep}
\title{Get the categories of ICD-9 and ICD-10 codes on diagnoses, the grouping rules are based on your standards.}
\usage{
IcdDxToCustomGrep(icdFile, icdColName, grepTable)
}
\arguments{
\item{icdFile}{ICD-9 and ICD-10 codes}

\item{icdColName}{A column for ICD of DxDataFile}

\item{grepTable}{Grouping rules of clustering the ICD is based on yourself! There are two column in the dataframe: Group, GrepICD#'}
}
\description{
This can be used to select the first diagnosis record
based on ICD code (grepl language) in DxDataFile,
return first diagnosis record based on factIcd Data
}
\examples{
icdFile <- data.frame(ICD = c("I95.0", "I952", "I110", "01091"), stringsAsFactors = FALSE)
grepTable <- data.frame(Group = c("Hypotension", "Hypertension"),
                            grep_pattern = c("^I95|^I952", "^I11"),
                            stringsAsFactors = FALSE)
IcdDxToCustomGrep(icdFile, ICD, grepTable)

}
