% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxToPhecode.R
\name{IcdDxToPhecode}
\alias{IcdDxToPhecode}
\title{Get the Phecode or description of ICD-9 diagnosis codes}
\source{
ICD-9-Phecode (version 1.2, 2015)

\url{https://phewascatalog.org/phecodes}
}
\usage{
IcdDxToPhecode(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, isDescription = TRUE)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{icd 10 using date}

\item{isDescription}{Phecode/ description for icd9. By default it is set to \code{True}.}
}
\description{
This can be used to group Phecode or description based on ICD-9 codes in clinical diagnostic data.
}
\details{
return Phecode or description based on ICD-9-CM codes
}
\examples{
head(sampleDxFile)
IcdDxToPhecode(sampleDxFile, ID, ICD, Date, "2015-10-01", FALSE)

}
