% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxToPheWAS.R
\name{IcdDxToPheWAS}
\alias{IcdDxToPheWAS}
\title{Get the PheWAS or description of ICD-9 diagnosis codes}
\source{
ICD-9-PheWAS (version 1.2, 2015)

\url{https://phewascatalog.org/PheWASs}
}
\usage{
IcdDxToPheWAS(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, isDescription = TRUE)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{icd 10 using date}

\item{isDescription}{PheWAS/ description for icd9. By default it is set to \code{True}.}
}
\description{
This can be used to group PheWAS or description based on ICD-9 codes in clinical diagnostic data.
}
\details{
return PheWAS or description based on ICD-9-CM codes
}
\examples{
head(sampleDxFile)
PheWAS <- IcdDxToPheWAS(sampleDxFile, ID, ICD, Date, "2015-10-01", FALSE)
head(PheWAS$groupedDT)
head(PheWAS$summarised_groupedDT)

}
