% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrDecimaltoShort.R
\name{IcdPrDecimaltoShort}
\alias{IcdPrDecimaltoShort}
\title{Convert ICD-9-PCS Codes From Decimal To Short Forms}
\usage{
IcdPrDecimaltoShort(icdList)
}
\arguments{
\item{icdList}{only ICD-9-PCS codes have decimal and short forms}
}
\description{
Return codes with short forms
}
