% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrShortToDecimal.R
\name{IcdPrShortToDecimal}
\alias{IcdPrShortToDecimal}
\title{Convert ICD-9-PCS Codes From Short To Decimal Forms}
\usage{
IcdPrShortToDecimal(icdList)
}
\arguments{
\item{icdList}{only ICD-9-PCS codes have decimal and short forms}
}
\description{
Return codes with Decimal forms
}
