% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrToCCS.R
\name{IcdPrToCCS}
\alias{IcdPrToCCS}
\title{Get the Clinical Classifications Software (CCS) categories and description for ICD-9 and ICD-10 codes on procedures.}
\source{
ICD-9-PCS CCS (2015)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Single_Level_CCS_2015.zip}

\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Multi_Level_CCS_2015.zip}

ICD-10-PCS CCS (2019)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs10/ccs_pr_icd10pcs_2019_1.zip}
}
\usage{
IcdPrToCCS(PrDataFile, idColName, icdColName, dateColName, icd10usingDate,
  isDescription = TRUE)
}
\arguments{
\item{PrDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of PrDataFile}

\item{icdColName}{A column for ICD of PrDataFile}

\item{dateColName}{A column for Date of PrDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{isDescription}{Clinical Classifications Software (CCS) single level categories/description for ICD-9 or ICD-10. By default it is set to \code{True}.}
}
\description{
Clinical Classifications Software (CCS) for ICD-9 and ICD-10 procedure codes in clinical diagnostic data is a procedure categorization scheme.
}
\details{
return Clinical Classifications Software (CCS) categories or description based on ICD-9 and ICD-10 codes
}
\examples{
head(samplePrFile)
IcdPrToCCS(samplePrFile, ID, ICD, Date, "2015-10-01", TRUE)

}
