% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrToCCSLvl.R
\name{IcdPrToCCSLvl}
\alias{IcdPrToCCSLvl}
\title{Get the Multi-level Clinical Classifications Software (CCS) categories and description for ICD-9 and ICD-10 codes on procedures.}
\source{
ICD-9-PCS CCS (2015)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Single_Level_CCS_2015.zip}

\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/Multi_Level_CCS_2015.zip}

ICD-10-PCS CCS (2019)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs10/ccs_pr_icd10pcs_2019_1.zip}
}
\usage{
IcdPrToCCSLvl(PrDataFile, idColName, icdColName, dateColName,
  icd10usingDate, CCSLevel = 1, CCSLvlLabel = TRUE)
}
\arguments{
\item{PrDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", "Date"}

\item{idColName}{A column for MemberID of PrDataFile}

\item{icdColName}{A column for ICD of PrDataFile}

\item{dateColName}{A column for Date of PrDataFile}

\item{icd10usingDate}{icd 10 using date}

\item{CCSLevel}{By default it is set to \code{1}.Clinical Classifications Software (CCS) multiple level}

\item{CCSLvlLabel}{Clinical Classifications Software (CCS) multiple level categories/description for icd9/10. By default it is set to \code{True}.}
}
\description{
Multi-leve Clinical Classifications Software (CCS) for ICD-9 and ICD-10 procedure codes in clinical procedures data is a procedure categorization scheme. Four levels exist in the multi-level procedure CCS for ICD-9-CM codes, and two levels exist in the multi-level procedure CCS for ICD-10-CM codes
}
\details{
return Multi-leve Clinical Classifications Software (CCS) categories or description based on ICD-9 and ICD-10 codes
}
\examples{
head(samplePrFile)
IcdPrToCCSLvl(samplePrFile, ID, ICD, Date, "2015-10-01", 2, TRUE)

}
