% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrToProcedureClass.R
\name{IcdPrToProcedureClass}
\alias{IcdPrToProcedureClass}
\title{Get the Procedure Class for ICD-9 and ICD-10 codes on procedures.}
\source{
ICD-9-Procedure Class (2015)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/procedure/pc2015.csv}

ICD-10-Procedure Class (2018)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/procedureicd10/procedure_icd10.jsp}
}
\usage{
IcdPrToProcedureClass(PrDataFile, idColName, icdColName, dateColName,
  icd10usingDate, isProcedureClassName = TRUE)
}
\arguments{
\item{PrDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of PrDataFile}

\item{icdColName}{A column for ICD of PrDataFile}

\item{dateColName}{A column for Date of PrDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{isProcedureClassName}{Procedure Class category/name for ICD-9 or ICD-10, default is True}
}
\description{
return Procedure Class categories based on ICD-9 and ICD-10 codes
}
\examples{
head(samplePrFile)
IcdPrToProcedureClass(samplePrFile, ID, ICD, Date, "2015-10-01", TRUE)

}
