% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientRecordDate.R
\name{patientRecordDate}
\alias{patientRecordDate}
\title{patients' first and last record date.}
\usage{
patientRecordDate(DxDataFile, idColName, icdColName, dateColName)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}
}
\description{
patients' first and last record date.
}
\examples{
head(sampleDxFile)
record <- patientRecordDate(sampleDxFile, ID, ICD, Date)
head(record)
}
