% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrToProcedureClass.R, R/fun_emr.R
\name{IcdPrToProcedureClass}
\alias{IcdPrToProcedureClass}
\alias{PC}
\title{Code classification for procedure code}
\source{
ICD-9-Procedure Class (2015)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/procedure/pc2015.csv}

ICD-10-Procedure Class (2018)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/procedureicd10/procedure_icd10.jsp}
}
\usage{
IcdPrToProcedureClass(PrDataFile, idColName, icdColName, dateColName,
  icd10usingDate, isDescription = TRUE)
}
\arguments{
\item{PrDataFile}{A file of clinical procedure data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A MemberID column of PrDataFile}

\item{icdColName}{A ICD column of PrDataFile}

\item{dateColName}{A Date column of PrDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{isDescription}{Category or description of standard classification methods for ICD procedure codes. By default, it set to \code{True} (standard description).}
}
\description{
The Procedure Class classification for ICD-9 and ICD-10 codes is a procedure categorization scheme that can employ in many types of projects analyzing data
}
\details{
Collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than are individual ICD procedure codes.
}
\examples{
# sample file for example

head(samplePrFile)

# Group procedure codes into procedure class classification

IcdPrToProcedureClass(samplePrFile, ID, ICD, Date, "2015-10-01", TRUE)

}
