% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{pcICD10}
\alias{pcICD10}
\title{Procedure Class for ICD-10-PCS}
\format{A data frame with 79758 rows and 4 variables}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/procedureicd10/pc_icd10pcs_2019_1.zip}
}
\usage{
data(pcICD10)
}
\description{
The Procedure Classes provide users an easy way to categorize procedure codes into one of four broad categories: Minor Diagnostic, Minor Therapeutic, Major Diagnostic, and Major Therapeutic.  (Updated for codes valid through FY 2019.)
}
\details{
\itemize{
  \item ICD. ICD-10-CM procedure codes
  \item ICD_DESCRIPTION. Code Description (word label for the code)
  \item PROCEDURE_CLASS. 1=minor diagnostic / 2=minor therapeutic / 3 = major diagnostic / 4=major therapeutic
  \item PROCEDURE_CLASS_NAME. minor diagnostic, minor therapeutic, major diagnostic, and major therapeutic
}
}
\seealso{
\code{\link{pcICD9}}
}
\keyword{datatable}
