% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{pcICD9}
\alias{pcICD9}
\title{Procedure Class for ICD-9-PCS}
\format{A data frame with 3948 rows and 4 variables}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/procedure/pc2015.csv}
}
\usage{
data(pcICD9)
}
\description{
The Procedure Classes provide users an easy way to categorize procedure codes into one of four broad categories: Minor Diagnostic, Minor Therapeutic, Major Diagnostic, and Major Therapeutic.  (Updated for codes valid through FY 2015.)
}
\details{
\itemize{
  \item ICD. ICD-9-CM procedure codes
  \item ICD_DESCRIPTION. Code Description (word label for the code)
  \item PROCEDURE_CLASS. 1=minor diagnostic / 2=minor therapeutic / 3 = major diagnostic / 4=major therapeutic
  \item PROCEDURE_CLASS_NAME. minor diagnostic, minor therapeutic, major diagnostic, and major therapeutic
}
}
\seealso{
\code{\link{pcICD10}}
}
\keyword{dataframe}
