% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{phecode_icd9_2}
\alias{phecode_icd9_2}
\title{Phecode for ICD-9}
\format{A data table with 15558 rows and 9 variables}
\source{
\url{https://phewascatalog.org/phecodes}
}
\usage{
data(phecode_icd9_2)
}
\description{
A datatable for ICD-9 Phecode mapping
}
\details{
\itemize{
  \item ICD. ICD-9
  \item ICDString. ICD-9 description
  \item PheCode. phecode for ICD-9
  \item PheCodeDescription. phecode description for ICD-9
  \item ExclPhecodes. exclude range of PheCode
  \item ExclPhenotypes. exclude range of PheCodeDescription
  \item Rollup. whether or not ICD9s mapped to this code also map to this code's parents
  \item Leaf. if phecode has parent phecode, leaf is 1.
  \item IgnoreBool. this icd9 code is ignored in analysis if this valus is 1.
}
}
\keyword{datatable}
