% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_errorICD.R
\name{plot_errorICD}
\alias{plot_errorICD}
\title{Pareto Plot of error ICD Codes}
\source{
\url{http://sape.inf.usi.ch/quick-reference/ggplot2/colour}
}
\usage{
plot_errorICD(errorFile, ICDVersion = all, wrongICDType = all,
  groupICD = FALSE, Others = TRUE, OthersThreshold)
}
\arguments{
\item{errorFile}{error file from ICD uniform function (`IcdDxDecimalToShort` or `IcdDxShortToDecimal`)}

\item{ICDVersion}{ICD version}

\item{wrongICDType}{Wrong ICD type}

\item{groupICD}{default is FALSE}

\item{Others}{default is TRUE}

\item{OthersThreshold}{Others threshold}
}
\description{
Pareto Plot of error ICD Codes
}
\examples{
head(sampleDxFile)
error <- IcdDxDecimalToShort(sampleDxFile, ICD, Date, "2015/10/01")
plot_errorICD(error$Error, ICDVersion = all,
                           wrongICDType = all,
                           groupICD = FALSE,
                           Others = FALSE,
                           OthersThreshold = 1)

}
