% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_groupedData.R
\name{plot_groupedData}
\alias{plot_groupedData}
\title{Plot of diagnostic category data}
\usage{
plot_groupedData(groupedDataWide, TopN = 10, limitPercentage = 0.01,
  pvalue = 0.05)
}
\arguments{
\item{groupedDataWide}{groupedData file from functions of code classification (four stPercentagegies)}

\item{TopN}{Default is Top "10"}

\item{limitPercentage}{the minimum  of Percentage Default is "0.01", in other words,the limit at same diagnostic category must have 1 percent patient in total patient.}

\item{pvalue}{p value of chisq.test}
}
\description{
Plot of diagnostic category data
}
\examples{
head(sampleDxFile)
groupedDataWide <- groupedDataLongToWide(sampleDxFile, ID, ICD, Date,
                                         icd10usingDate = "2015-10-01",
                                         groupDataType = elix)
plot1 <- plot_groupedData(groupedDataWide = groupedDataWide,
                          TopN = 10,
                          limitPercentage = 0.01)

selectedCaseFile <- selectCases(DxDataFile = sampleDxFile,
                                idColName = ID,
                                icdColName = ICD,
                                dateColName = Date,
                                icd10usingDate = "2015/10/01",
                                groupDataType = ccslvl2,
                                caseCondition = "Diseases of the urinary system",
                                caseCount = 1)
groupedDataWide <- groupedDataLongToWide(sampleDxFile, ID, ICD, Date,
                                         icd10usingDate = "2015-10-01",
                                         groupDataType = elix,
                                         selectedCaseFile = selectedCaseFile)
plot2 <- plot_groupedData(groupedDataWide = groupedDataWide,
                          TopN = 10,
                          limitPercentage = 0.01,
                          pvalue = 0.05)
plot1
plot2

}
