% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/plot_errorICD.R
\name{plotError}
\alias{plotError}
\alias{plot_errorICD}
\title{Plot for error ICD list}
\usage{
plot_errorICD(errorFile, icdVersion = all, wrongICDType = all,
  groupICD = FALSE, others = TRUE, topN = 10)
}
\arguments{
\item{errorFile}{Error file from ICD uniform function (`icdDxDecimalToShort` or `icdDxShortToDecimal`)}

\item{icdVersion}{ICD version: ICD9 (\code{'9'}), ICD10 (\code{'10'}, and all version \code{'all'}}

\item{wrongICDType}{Wrong ICD type: wrong version (\code{'version'}), wrong format (\code{'format'}, and both wrong type \code{'all'}}

\item{groupICD}{Only ICD-9 codes can group, because ICD 10 already has unique alphanumeric codes to identify known diseases. Default is FALSE}

\item{others}{Default is TRUE}

\item{topN}{Default is Top "10"}
}
\value{
A Pareto plot and a \code{data.table} for error codes.
}
\description{
Pareto chart of error ICD list
}
\details{
Through first phase function, code standardization, which generates a data of diagnosis codes with potential error. The Pareto chart includes bar plot and line chart to visualize individual possible error ICD codes represented in descending order and cumulative total.
}
\examples{
# sample file for example
head(sampleDxFile)

# Data of diagnosis codes with potential error

error <- icdDxDecimalToShort(sampleDxFile, ICD, Date, icd10usingDate = "2015/10/01")

# Plot of top 5 common error ICD-9 codes and a list of the detail of error ICD codes

plot_errorICD(errorFile = error$Error,
              icdVersion = 9,
              wrongICDType = all,
              groupICD = TRUE,
              others = TRUE,
              topN = 5)

# Plot of top 10 common error ICD codes and a list of the detail of error ICD codes

plot_errorICD(errorFile = error$Error,
              icdVersion = all,
              wrongICDType = all,
              groupICD = FALSE,
              others = TRUE)

}
\seealso{
other plot function: \code{\link{plotGroupedData}}
}
