% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/plot_groupedData.R
\name{plotGroupedData}
\alias{plotGroupedData}
\alias{plot_groupedData}
\title{Plot of diagnostic categories}
\usage{
plot_groupedData(groupedDataWide, topN = 10, limitFreq = 0.01,
  pvalue = 0.05)
}
\arguments{
\item{groupedDataWide}{GroupedData file from functions of code classification (four stPercentagegies)}

\item{limitFreq}{The minimum frequency set to "0.01"; In other words, the limit at the same diagnostic category must have 1 percent patient in the total patient.}

\item{pvalue}{p value of chisq.test}

\item{TopN}{Default is Top "10"}
}
\value{
A histogram plot and a \code{data.table} for classifying data.
}
\description{
Histogram plot of diagnostic categories
}
\details{
An overview of grouping category of the diagnostic code summarizes the percentage of result in histogram plot. User can observe the number of diagnostic categories in their dataset through this function. This function also verifies by Chi-square test and Fisher’s exact test statistical significantly different diagnostic categories between case and control. The default level of significance considered at 5% (p = 0.05).
}
\examples{
# sample file for example
head(sampleDxFile)

# Convert long format of grouped data into wide binary format

groupedDataWide <- groupedDataLongToWide(sampleDxFile, ID, ICD, Date,
                                         icd10usingDate = "2015-10-01",
                                         groupDataType = elix,
                                         isDescription = FALSE)

# plot of top 10 common grouped categories and a list of the detail of grouped categories

plot1 <- plot_groupedData(groupedDataWide = groupedDataWide,
                          topN = 10,
                          limitFreq = 0.01)
plot1

# Select case with "Diseases of the urinary system" by level 2 of CCS classification

selectedCaseFile <- selectCases(dxDataFile = sampleDxFile,
                                idColName = ID,
                                icdColName = ICD,
                                dateColName = Date,
                                icd10usingDate = "2015/10/01",
                                groupDataType = ccslvl2,
                                caseCondition = "Diseases of the urinary system",
                                caseCount = 1)

# Convert the long format of grouped data into a wide binary format with selected case

groupedDataWide <- groupedDataLongToWide(sampleDxFile, ID, ICD, Date,
                                         icd10usingDate = "2015-10-01",
                                         groupDataType = icd,
                                         isDescription = FALSE,
                                         selectedCaseFile = selectedCaseFile)

# plot of top 10 common grouped categories and a list of the detail of grouped categories

plot2 <- plot_groupedData(groupedDataWide = groupedDataWide,
                          topN = 10,
                          limitFreq = 0.01,
                          pvalue = 0.05)
plot2
}
\seealso{
other plot function: \code{\link{plotError}}
}
