% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_emr.R
\docType{data}
\name{Pr_dataset_classification}
\alias{Pr_dataset_classification}
\alias{ccsPrICD9}
\alias{ccsPrICD10}
\alias{pcICD9}
\alias{pcICD10}
\title{Datasets for code classification}
\format{These datatable with ICD procedure codes and the specific Standardization categories.}
\usage{
ccsPrICD9

ccsPrICD10

pcICD9

pcICD10
}
\description{
These datasets are used for procedure code transformation.
}
\details{
These datasets are used for \code{IcdPrToCCS}, \code{IcdPrToCCSLvl}, and \code{IcdPrToProcedureClass}
}
\keyword{datasets}
