% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{prICD10}
\alias{prICD10}
\title{ICD-10-PCS}
\format{A data frame with 78880 rows and 2 variables}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD10/2019-ICD-10-PCS.html}
}
\usage{
data(prICD10)
}
\description{
A datatable for ICD-10-PCS, containing ICD-10-PCS code and ICD description (Updated for codes valid through FY 2019).
}
\details{
\itemize{
  \item ICD. ICD-10-PCS
  \item ICD_DESCRIPTION. ICD-10-PCS description
}
}
\seealso{
\code{\link{prICD9}}
}
\keyword{datatable}
