% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{prICD9}
\alias{prICD9}
\title{ICD-9-PCS}
\format{A data frame with 3882 rows and 2 variables}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}
}
\usage{
data(prICD9)
}
\description{
A dataframe for ICD-9-PCS, containing ICD-9-PCS code and ICD description(Updated for codes valid through FY 2014.)
}
\details{
\itemize{
  \item ICD. ICD-9-PCS
  \item ICD_DESCRIPTION. ICD-9-PCS description
}
}
\seealso{
\code{\link{prICD10}}
}
\keyword{dataframe}
