% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrDecimaltoShort.R, R/IcdPrShortToDecimal.R,
%   R/fun_emr.R
\name{IcdPrDecimalToShort}
\alias{IcdPrDecimalToShort}
\alias{IcdPrShortToDecimal}
\alias{PrUniform}
\title{Code format transformation}
\source{
\url{https://www.findacode.com/search/search.php}

\url{https://www.cms.gov/Medicare/Quality-Initiatives-Patient-Assessment-Instruments/HospitalQualityInits/Downloads/HospitalAppendix_F.pdf}

\url{https://www.cms.gov/Medicare/Coding/ICD10/2019-ICD-10-PCS.html}
}
\usage{
IcdPrDecimalToShort(PrDataFile, icdColName, dateColName, icd10usingDate)

IcdPrShortToDecimal(PrDataFile, icdColName, dateColName, icd10usingDate)
}
\arguments{
\item{PrDataFile}{A file of clinical procedure data with at least 2 columns: "ICD" and "Date"}

\item{icdColName}{A ICD column of PrDataFile}

\item{dateColName}{A Date column of PrDataFile}

\item{icd10usingDate}{ICD-10 using date}
}
\description{
Converts the ICD procedure codes to a uniform format
}
\details{
These transformative function does not only convert the ICD to uniform format code but also check the potential coding error of the ICD codes’ format or version.
}
\examples{
# sample file for example

head(samplePrFile)

# convert the procedure codes to the short format

IcdPrDecimalToShort(samplePrFile,ICD,Date,"2015/10/01")

# convert the procedure codes to the decimal format

IcdPrShortToDecimal(samplePrFile,ICD,Date,"2015/10/01")
}
