% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_emr.R
\docType{data}
\name{sampleDxFile}
\alias{sampleDxFile}
\title{Sample file for demo with diagnostic codes}
\format{
A sample data table with 300 rows and 3 variables
}
\description{
Demo for data integration, data wrangling and visualization.
}
