% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_emr.R
\docType{data}
\name{sampleMIMICIII}
\alias{sampleMIMICIII}
\title{Sample file for patients with patent ductus arteriosus (PDA) from MIMIC-III.}
\format{
A data table with 45674 rows and 8 variables
}
\source{
Pollard, T. J., & Johnson, A. E. W. The MIMIC-III Clinical Database {http://dx.doi.org/10.13026/C2XW26} (2016)
}
\description{
Demo for data analysis.
}
