% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_emr.R
\docType{data}
\name{samplePrFile}
\alias{samplePrFile}
\title{Sample file for demo with procedure codes}
\format{A data table with 170 rows and 3 variables}
\description{
Demo for code standardization, data integration.
}
