% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/selectCases.R
\name{selectCase}
\alias{selectCase}
\alias{selectCases}
\title{Data integration for case selection}
\usage{
selectCases(DxDataFile, idColName, icdColName, dateColName, icd10usingDate,
  groupDataType = ccs, CustomGroupingTable, isDescription = TRUE,
  caseCondition, caseCount, PeriodRange = c(30, 365),
  CaseName = "Selected")
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of DxDataFile.}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{groupDataType}{Four Stratified methods can chosen: CCS (\code{'ccs'}), CCS levels (\code{'ccslvl1'}, \code{'ccslvl2'}, \code{'ccslvl3'}, \code{'ccslvl4'}), PheWAS (\code{'PheWAS'}), comorbidities (\code{'ahrq'},\code{'charlson'}, \code{'elix'}), precise or fuzzy customized  method (\code{'customGrepIcdGroup'}, \code{'customIcdGroup'}). Change it to any of the other possible variables, default it set to \code{"ccs"}.}

\item{CustomGroupingTable}{Define the grouping categories. `IcdDxToCustom` needs a dataset with two columns: "Group", "ICD"; "Group" defines one or more disease categories. "ICD" defines a list of disease-related ICD codes. `IcdDxToCustomGrep`needs a dataset with two columns: "Group", "grepIcd"; "Group" defines one or more disease categories. "grepICD" defines disease-related ICD code character strings containing a regular expression.}

\item{isDescription}{Category or description of standard classification methods for ICD diagnostic codes. By default, it set to \code{True} (standard description).}

\item{caseCondition}{Certain diseases of standard groups.}

\item{caseCount}{Minimum number of diagnoses.}

\item{PeriodRange}{Determine days of interest for performing the case selection. By default, it set from 30 to 365 days.}

\item{CaseName}{Name of selected cases. By default, it set to \code{"selected"}.}
}
\value{
A new \code{data.table} based on standard classification dataset with a new column: \code{selectedCase}
}
\description{
This query function can select the cases matching defined conditions for analyses.
}
\details{
User can select cases by diagnostic categories, such as CCS category, ICD codes, etc. The function also provides the options to set the minimum number of diagnoses within a specific duration. The output dataset can be passed to `groupedDataLongToWide` to create tables in wide format for statistical analytic usage.
}
\examples{
# sample file for example

head(sampleDxFile)

#select case with "Diseases of the urinary system" by level 2 of CCS classification

selectCases(DxDataFile = sampleDxFile,
            ID, ICD, Date,
            groupDataType = ccslvl2,
            icd10usingDate = "2015/10/01",
            caseCondition = "Diseases of the urinary system",
            caseCount = 1)
}
\seealso{
Other data integration functions: \code{\link{DataSplit}}, \code{\link{recordPeriod}}, \code{\link{era}}
}
