% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/selectCases.R
\name{selectCases}
\alias{selectCases}
\title{Data integration for case selection}
\usage{
selectCases(dxDataFile, idColName, icdColName, dateColName,
  icdVerColName = NULL, icd10usingDate = NULL, groupDataType = CCS,
  customGroupingTable, isDescription = TRUE, caseCondition, caseCount,
  periodRange = c(30, 365), caseName = "Selected")
}
\arguments{
\item{dxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of dxDataFile}

\item{icdColName}{A column for ICD of dxDataFile}

\item{dateColName}{A column for Date of dxDataFile}

\item{icdVerColName}{A column for ICD Version used of dxDataFile. If this argument is provided, then icd10usingDate is not needed.}

\item{icd10usingDate}{The date that ICD-10 was started using}

\item{groupDataType}{Four Stratified methods can be chosen: CCS (\code{'ccs'}), CCS levels (\code{'ccslvl1'}, \code{'ccslvl2'}, \code{'ccslvl3'}, \code{'ccslvl4'}), PheWAS (\code{'PheWAS'}), comorbidities (\code{'ahrq'},\code{'charlson'}, \code{'elix'}), precise or fuzzy customized  method (\code{'customGrepIcdGroup'}, \code{'customIcdGroup'}). Change it to any of the other possible variables, default it set to \code{"ccs"}.}

\item{customGroupingTable}{Define the grouping categories. `icdDxToCustom` needs a dataset with two columns: "Group", "ICD"; "Group" defines one or more disease categories and "ICD" defines a list of disease-related ICD codes. `icdDxToCustomGrep`needs a dataset with two columns: "Group", "grepIcd"; "Group" defines one or more disease categories and "grepICD" defines disease-related ICD code character strings containing a regular expression.}

\item{isDescription}{Category or description of standard classification methods for ICD diagnostic codes. The default value is \code{True} (standard description).}

\item{caseCondition}{Certain diseases' conditions that interested}

\item{caseCount}{Total numbers of selected diagnoses}

\item{periodRange}{Determine days of interest for performing the case selection. By default, the range is set between 30 to 365 days.}

\item{caseName}{Name of selected cases. By default, it is set to \code{"selected"}.}
}
\value{
A new \code{data.table} based on standard classification dataset with a new column: \code{selectedCase}
}
\description{
This query function can select the cases matching defined conditions for analyses.
}
\details{
User can select cases by diagnostic categories, such as CCS category, ICD codes, etc. The function also provides the options to set the minimum number of diagnoses within a specific duration. The output dataset can be passed to `groupedDataLongToWide` to create tables in wide format for statistical analytic usage.
}
\examples{
# sample file for example

head(sampleDxFile)

# select cases with "Diseases of the urinary system" by level 2 of CCS classification

selectCases(dxDataFile = sampleDxFile,
            idColName = ID,
            icdColName = ICD,
            dateColName = Date,
            icdVerColName =  NULL,
            icd10usingDate = "2015/10/01",
            groupDataType = ccslvl2,
            caseCondition = "Diseases of the urinary system",
            caseCount = 1,
            periodRange = c(30,180),
            caseName = "Selected")

# select cases with ICD codes started with "585" and "586"

selectCases(dxDataFile = sampleDxFile,
            idColName = ID,
            icdColName = ICD,
            dateColName = Date,
            icdVerColName =  NULL,
            icd10usingDate = "2015/10/01",
            groupDataType = ICD,
            caseCondition = "585|586",
            caseCount = 3,
            periodRange = c(30,180),
            caseName = "selected")
}
\seealso{
Other data integration functions: \code{\link{dataSplit}}, \code{\link{recordPeriod}}, \code{\link{era}}
}
