% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDataByDate.R
\name{splitDataByDate}
\alias{splitDataByDate}
\title{Select cases based on ICD code and the number of ICD codes}
\usage{
splitDataByDate(DxDataFile, idColName, icdColName, dateColName, IndexDate,
  window = 30)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{IndexDate}{An exact date of diagnosis for a period of observation.}

\item{window}{Length of condition era. By default it is set to \code{30}.}
}
\description{
This can be used to select qualified cases from factIcd data
based on the ICD code searching criteria and number of ICD code
per patients in the inout factIcd dataset.
Return qualified Members' data
}
\examples{
head(sampleDxFile)
splitDataByDate(sampleDxFile, ID, ICD, Date,"2015-10-01",30)

}
