% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{testDxFile}
\alias{testDxFile}
\title{Test File of ICD-CM codes for grouping method}
\format{A data frame with 4002 rows and 3 variables}
\usage{
data(testDxFile)
}
\description{
A test file for grouping method (CCS, CCSLvl, phecode, comorbidity...) and calculating condition era
}
\details{
\itemize{
  \item ID  MemberID
  \item ICD ICD-9 and ICD-10-CM codes
  \item Date diagnosis date
}
}
\keyword{dataframe}
