% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R
\name{common_PrArg}
\alias{common_PrArg}
\title{Common argument of procedure functions}
\arguments{
\item{prDataFile}{A data frame object of clinical procedure data with at least 3 columns: ID, ICD, and Date. As for date column, the data format should be YYYY/MM/DD or YYYY-MM-DD.}

\item{idColName}{Column name of ID column in prDataFile. Data type of this argumant should be string without quotation marks.}

\item{icdColName}{Column name of ICD column in prDataFile. Data type of this argumant should be string without quotation marks.}

\item{dateColName}{Column name of date column in prDataFile (with date in YYYY/MM/DD or YYYY-MM-DD format). Data type of this argumant should be string without quotation marks.}

\item{icd10usingDate}{ICD-10 using date}

\item{isDescription}{Category or description of standard classification methods for ICD procedure codes. By default, it set to \code{True} (standard description).}
}
\description{
Common argument of procedure functions
}
\details{
Common argument of procedure functions
}
