% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIcdDecimaltoShort.R
\name{convertIcdDecimaltoShort}
\alias{convertIcdDecimaltoShort}
\title{Get the Clinical Classifications Software (CCS) single and multiple category/ description}
\usage{
convertIcdDecimaltoShort(icdList)
}
\arguments{
\item{icdList}{An icd code list}
}
\description{
This can be used to select the Clinical Classifications Software (CCS) single and multiple category/ description
based on ICD code in clinical diagnostic data,
return Clinical Classifications Software (CCS) single and multiple category/ description based on ICD-9 and ICD-10
}
