% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIcdDecimaltoShort.R
\name{convertIcdDecimaltoShort}
\alias{convertIcdDecimaltoShort}
\title{Convert ICD Codes From Decimal To Short Forms}
\usage{
convertIcdDecimaltoShort(icdList)
}
\arguments{
\item{icdList}{ICD codes}
}
\description{
Convert codes between short and decimal forms
}
