% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIcdShortToDecimal.R
\name{convertIcdShortToDecimal}
\alias{convertIcdShortToDecimal}
\title{Get the PheWAS Phecode/ description}
\usage{
convertIcdShortToDecimal(icdList)
}
\arguments{
\item{icdList}{An icd code list}
}
\description{
This can be used to select the PheWAS Phecode/ description
based on ICD code in clinical diagnostic data,
return CCS single and multiple category/ description based on ICD
}
