% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIcdShortToDecimal.R
\name{convertIcdShortToDecimal}
\alias{convertIcdShortToDecimal}
\title{Convert ICD Codes From Short To Decimal Forms}
\usage{
convertIcdShortToDecimal(icdList)
}
\arguments{
\item{icdList}{ICD codes}
}
\description{
Convert codes between short and decimal forms
}
