% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/splitDataByDate.R
\name{dataSplit}
\alias{dataSplit}
\alias{splitDataByDate}
\title{Data integration for data split}
\usage{
splitDataByDate(dxDataFile, idColName, icdColName, dateColName,
  indexDateFile, gap = 30)
}
\arguments{
\item{dxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of dxDataFile.}

\item{icdColName}{A column for ICD of dxDataFile}

\item{dateColName}{A column for Date of dxDataFile}

\item{indexDateFile}{Diagnostic dates for each patient in an observed period.}

\item{Gap}{Gap length of the window. Default set to \code{30}.}
}
\value{
A new \code{data.table} based on \code{dxDataFile} and classified by \code{indexDateFile} for each patient.
}
\description{
Splitting data by the date of the clinical event and shows the data recorded before or after the clinical event and the window counts, which is the gap between the record date and index date.
}
\details{
In most condition, users need to extract data by a specific clinical event (e.g., first diagnosis dates of chronic diseases). Users can define a table of clinical index dates of each patient. The date can be generated by `selectCases` function or first/last admission date by `patientRecordDate` function.
}
\examples{
# sample file for example

SampleforCertainPatient <- sampleDxFile[grepl("A0|B0|C0|D0",ID),]

head(SampleforCertainPatient)

# Defined index date of patient A0,B0,C0 and D0

indexDateTable <- data.frame(ID = c("A0","B0","C0","D0"),
                             indexDate = c("2023-08-12", "2024-02-12",
                                           "2015-12-05", "2017-01-29"),
                             stringsAsFactors = FALSE)
indexDateTable
# Split data by index date for each patient

splitedData <- splitDataByDate(SampleforCertainPatient, ID, ICD, Date,
                               indexDateFile = indexDateTable,
                               gap = 30)
splitedData[15:19,]
}
\seealso{
Other data integration functions: \code{\link{selectCases}}, \code{\link{recordPeriod}}, \code{\link{era}}
}
