% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/groupedDataLongToWide.R
\name{dataWide}
\alias{dataWide}
\alias{groupedDataLongToWide}
\title{Data format transformation}
\usage{
groupedDataLongToWide(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, groupDataType = ccs, CustomGroupingTable,
  isDescription = TRUE, numericOrBinary = B, selectedCaseFile = NULL)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of DxDataFile.}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{groupDataType}{Four Stratified methods can chosen: CCS (\code{'ccs'}), CCS levels (\code{'ccslvl1'}, \code{'ccslvl2'}, \code{'ccslvl3'}, \code{'ccslvl4'}), PheWAS (\code{'PheWAS'}), comorbidities (\code{'ahrq'},\code{'charlson'}, \code{'elix'}), precise or fuzzy customized  method (\code{'customGrepIcdGroup'}, \code{'customIcdGroup'}). Change it to any of the other possible variables, default it set to \code{"ccs"}.}

\item{CustomGroupingTable}{Define the grouping categories. `IcdDxToCustom` needs a dataset with two columns: "Group", "ICD"; "Group" defines one or more disease categories. "ICD" defines a list of disease-related ICD codes. `IcdDxToCustomGrep`needs a dataset with two columns: "Group", "grepIcd"; "Group" defines one or more disease categories. "grepICD" defines disease-related ICD code character strings containing a regular expression.}

\item{isDescription}{Category or description of standard classification methods for ICD diagnostic codes. By default, it set to \code{True} (standard description).}

\item{numericOrBinary}{Members have same diagnostic categories, type `N` or `B`, default is Binary \code{'B'}.}

\item{selectedCaseFile}{Table for selectedCases. Default is \code{'NULL'}}
}
\value{
A new \code{data.table} based on classifying \code{DxDataFile} and converted the dataset into a wide format dataset.
}
\description{
Convert the long format of grouped data into a wide format which is fit to others analytical and plotting packages.
}
\details{
The output dataset of this function can be numeric or binary wide format
}
\examples{
# sample file for example

head(sampleDxFile)

# Select case with "Diseases of the urinary system" by level 2 of CCS classification

selectedCaseFile <- selectCases(sampleDxFile, ID, ICD, Date,
                                icd10usingDate = "2015/10/01",
                                groupDataType = ccslvl2,
                                caseCondition = "Diseases of the urinary system",
                                caseCount = 1)

# Convert long format grouped data into a wide numeric format with selected case

groupedData_Wide <- groupedDataLongToWide(sampleDxFile, ID, ICD, Date,
                                          "2015-10-01", elix,
                                          numericOrBinary = N,
                                          isDescription = FALSE,
                                          selectedCaseFile = selectedCaseFile)
head(groupedData_Wide)
}
