% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/groupedDataLongToWide.R
\name{dataWide}
\alias{dataWide}
\alias{groupedDataLongToWide}
\title{Data format transformation}
\usage{
groupedDataLongToWide(
  dxDataFile,
  idColName,
  categoryColName,
  dateColName,
  reDup = TRUE,
  numericOrBinary = B,
  count = 1,
  selectedCaseFile = NULL
)
}
\arguments{
\item{dxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of dxDataFile.}

\item{dateColName}{A column for Date of dxDataFile}

\item{numericOrBinary}{Members have same diagnostic categories, type `N` or `B`, default is Binary \code{'B'}.}

\item{selectedCaseFile}{Table for selectedCases. Default is \code{'NULL'}}
}
\value{
A new \code{data.table} based on classifying \code{dxDataFile} and converted the dataset into a wide format dataset.
}
\description{
Convert the long format of grouped data into a wide format which is fit to others analytical and plotting packages.
}
\details{
The output dataset of this function can be numeric or binary wide format
}
\examples{

# Create a grouped data

ELIX <- icdDxToComorbid(dxDataFile = sampleDxFile,
                       idColName = ID,
                       icdColName = ICD,
                       dateColName = Date,
                       icd10usingDate = "2015/10/01",
                       comorbidMethod = elix)

head(ELIX$groupedDT)

# Select case with "Diseases of the urinary system" by level 2 of CCS classification

selectedCaseFile <- selectCases(dxDataFile = sampleDxFile,
                                idColName = ID,
                                icdColName = ICD,
                                dateColName = Date,
                                icdVerColName = NULL,
                                icd10usingDate = "2015/10/01",
                                groupDataType = ccslvl2,
                                caseCondition = "Diseases of the urinary system",
                                caseCount = 1)

# Convert the long format of grouped data into a wide binary format with selected case

groupedDataWide <- groupedDataLongToWide(ELIX$groupedDT,
                                         idColName = ID,
                                         categoryColName = Comorbidity,
                                         dateColName = Date,
                                         selectedCaseFile = selectedCaseFile)
groupedDataWide
}
