% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_emr.R
\docType{data}
\name{Dx_dataset_classification}
\alias{Dx_dataset_classification}
\alias{ccsDxICD9}
\alias{ccsDxICD10}
\alias{phecode_icd9_2}
\alias{icd9_ahrq}
\alias{icd10_ahrq}
\alias{icd9_charlson}
\alias{icd10_charlson}
\alias{icd9_elix}
\alias{icd10_elix}
\title{Datasets for code classification}
\format{These datatable with ICD diagnostic codes and the specific Standardization categories.}
\usage{
ccsDxICD9

ccsDxICD10

phecode_icd9_2

icd9_ahrq

icd10_ahrq

icd9_charlson

icd10_charlson

icd9_elix

icd10_elix
}
\description{
These datasets are used for diagnostic code transformation.
}
\details{
These datasets are used for \code{icdDxToCCS}, \code{icdDxToCCSLvl}, \code{icdDxToPheWAS}, and \code{icdDxToComorbid}
}
\keyword{datasets}
