% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxToCCS.R, R/IcdDxToCCSLvl.R, R/fun_emr.R
\name{IcdDxToCCS}
\alias{IcdDxToCCS}
\alias{IcdDxToCCSLvl}
\alias{DxCCS}
\title{Code classification for CCS}
\usage{
IcdDxToCCS(DxDataFile, idColName, icdColName, dateColName, icd10usingDate,
  isDescription = TRUE)

IcdDxToCCSLvl(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, CCSLevel = 1, isDescription = TRUE)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of DxDataFile.}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{isDescription}{Category or description of standard classification methods for ICD diagnostic codes. By default, it set to \code{True} (standard description).}

\item{CCSLevel}{By default it set to \code{1}. CCS multiple level:1~4, CCS for ICD-10-CM has only 1~2 multiple levels.}
}
\value{
Three new \code{data.table}s. 1) \code{groupedDT}: Based on \code{DxDataFile} with two new columns for uniform format diagnostic codes and classified standard categories. 2) \code{summarised_groupedDT}: Summarized the dataset  \code{groupedDT} and sorted by memberID. 3) \code{Error}: Potential error codes from \code{\link{DxUniform}}.
}
\description{
The CCS classification for ICD-9 and ICD-10 codes is a diagnostic categorization scheme that can employ in many types of projects analyzing data on diagnoses.
}
\details{
These CCS functions (IcdDxToCCS and IcdDxToCCSLvl) collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than are individual ICD diagnostic codes.
}
\examples{
# sample file for example

head(sampleDxFile)

# Group diagnostic codes into single level of CCS classification

IcdDxToCCS(sampleDxFile, ID, ICD, Date, "2015-10-01", TRUE)

# Group diagnostic codes into multiple levels of CCS classification

IcdDxToCCSLvl(sampleDxFile, ID, ICD, Date, "2015-10-01", 2, TRUE)
}
\seealso{
Other code classification functions: \code{\link{DxPheWAS}}, \code{\link{DxCustom}}, \code{\link{DxComorbid}}
}
