% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/icdDxToCCS.R, R/icdDxToCCSLvl.R
\name{dxCCS}
\alias{dxCCS}
\alias{icdDxToCCS}
\alias{icdDxToCCSLvl}
\title{Code classification for CCS}
\usage{
icdDxToCCS(dxDataFile, idColName, icdColName, dateColName,
  icdVerColName = NULL, icd10usingDate = NULL, isDescription = TRUE)

icdDxToCCSLvl(dxDataFile, idColName, icdColName, dateColName,
  icdVerColName = NULL, icd10usingDate = NULL, CCSLevel = 1,
  isDescription = TRUE)
}
\arguments{
\item{dxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of dxDataFile.}

\item{icdColName}{A column for ICD of dxDataFile}

\item{dateColName}{A column for Date of dxDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{isDescription}{Category or description of standard classification methods for ICD diagnostic codes. By default, it set to \code{True} (standard description).}

\item{CCSLevel}{By default it set to \code{1}. CCS multiple level:1~4, CCS for ICD-10-CM has only 1~2 multiple levels.}
}
\value{
Three new \code{data.table}s. 1) \code{groupedDT}: Based on \code{dxDataFile} with two new columns for uniform format diagnostic codes and classified standard categories. 2) \code{summarised_groupedDT}: Summarized the dataset  \code{groupedDT} and sorted by memberID. 3) \code{Error}: Potential error codes from \code{\link{dxUniform}}.
}
\description{
The CCS classification for ICD-9 and ICD-10 codes is a diagnostic categorization scheme that can employ in many types of projects analyzing data on diagnoses.
}
\details{
These CCS functions (icdDxToCCS and icdDxToCCSLvl) collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than are individual ICD diagnostic codes.
}
\examples{
# sample file for example

head(sampleDxFile)

# Group diagnostic codes into single level of CCS classification

icdDxToCCS(sampleDxFile, ID, ICD, Date, "2015-10-01", TRUE)

# Group diagnostic codes into multiple levels of CCS classification

icdDxToCCSLvl(sampleDxFile, ID, ICD, Date, "2015-10-01", 2, TRUE)
}
\seealso{
Other code classification functions: \code{\link{dxPheWAS}}, \code{\link{dxCustom}}, \code{\link{dxComorbid}}
}
