% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxToComorbid.R, R/fun_emr.R
\name{IcdDxToComorbid}
\alias{IcdDxToComorbid}
\alias{DxComorbid}
\title{Code classification for Comorbidity}
\usage{
IcdDxToComorbid(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, comorbidMethod, isDescription = FALSE)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of DxDataFile.}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{comorbidMethod}{Three comorbidity methods: AHRQ, Charlson and Elixhauser Comorbidity. Change it to any of the other possible variables (\code{ahrq},\code{charlson}, and \code{elix}).}

\item{isDescription}{Category or description of standard classification methods for ICD diagnostic codes. By default it is set to \code{FALSE} ( Comorbidity categories).}
}
\value{
Three new \code{data.table}s. 1) \code{groupedDT}: Based on \code{DxDataFile} with two new columns for uniform format diagnostic codes and classified standard categories. 2) \code{summarised_groupedDT}: Summarized the dataset  \code{groupedDT} and sorted by memberID. 3) \code{Error}: Potential error codes from \code{\link{DxUniform}}.
}
\description{
The comorbidities classification (AHRQ, Charlson ,and Elixhauser Comorbidity) for ICD diagnostic codes is a diagnostic categorization scheme that can employ in many types of projects analyzing data on diagnoses.
}
\details{
Collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than are individual ICD diagnostic codes.
}
\examples{
# sample file for example

head(sampleDxFile)

# Group diagnostic codes into charlson comorbidity categories

IcdDxToComorbid(sampleDxFile, ID, ICD, Date, "2015-10-01", charlson)
}
\seealso{
Other code classification functions: \code{\link{DxPheWAS}}, \code{\link{DxCustom}}, \code{\link{DxCCS}}
}
