% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/icdDxToCustom.R,
%   R/icdDxToCustomGrep.R
\name{dxCustom}
\alias{dxCustom}
\alias{icdDxToCustom}
\alias{icdDxToCustomGrep}
\title{Code classification for customized group}
\usage{
icdDxToCustom(dxDataFile, idColName, icdColName, dateColName,
  customGroupingTable)

icdDxToCustomGrep(dxDataFile, idColName, icdColName, dateColName,
  customGroupingTable)
}
\arguments{
\item{dxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of dxDataFile.}

\item{icdColName}{A column for ICD of dxDataFile}

\item{dateColName}{A column for Date of dxDataFile}

\item{customGroupingTable}{Define the grouping categories. `icdDxToCustom` needs a dataset with two columns: "Group", "ICD"; "Group" defines one or more disease categories. "ICD" defines a list of disease-related ICD codes. `icdDxToCustomGrep`needs a dataset with two columns: "Group", "grepIcd"; "Group" defines one or more disease categories. "grepICD" defines disease-related ICD code character strings containing a regular expression.}
}
\value{
Two new \code{data.table}s. 1) \code{groupedDT}: Based on \code{dxDataFile} with two new columns for uniform format diagnostic codes and classified standard categories. 2) \code{summarised_groupedDT}: Summarized the dataset  \code{groupedDT} and sorted by memberID.
}
\description{
Researches can define the grouping categories and therefore have more flexible for grouping ICD diagnostic codes.
}
\details{
There are two functions for customized defined grouping method, the customized category grouping is based on precise (`icdDxToCustom`) and fuzzy (`icdDxToCustomGrep`) grouping method, respectively.
}
\examples{
# sample file for example

head(sampleDxFile)

# Group diagnostic codes into "Chronic kidney disease" with precise grouping method

groupingTable <- data.frame(Group = rep("Chronic kidney disease",6),
                            ICD = c("N181","5853","5854","5855","5856","5859"),
                            stringsAsFactors = FALSE)

icdDxToCustom(sampleDxFile, ID, ICD, Date, CustomGroupingTable = groupingTable)

# Group diagnostic codes into "Chronic kidney disease" with fuzzy grouping method

grepTable <- data.frame(Group = "Chronic kidney disease",
                        grepIcd = "^585|^N18",
                        stringsAsFactors = FALSE)

icdDxToCustomGrep(sampleDxFile,ID, ICD, Date,CustomGroupingTable = grepTable)

}
\seealso{
Other code classification functions: \code{\link{dxPheWAS}}, \code{\link{dxComorbid}}, \code{\link{DxCCS}}
}
