% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxDecimalToShort.R, R/IcdDxShortToDecimal.R,
%   R/fun_emr.R
\name{IcdDxDecimalToShort}
\alias{IcdDxDecimalToShort}
\alias{IcdDxShortToDecimal}
\alias{DxUniform}
\title{Code format transformation}
\usage{
IcdDxDecimalToShort(DxDataFile, icdColName, dateColName, icd10usingDate)

IcdDxShortToDecimal(DxDataFile, icdColName, dateColName, icd10usingDate)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 2 columns: "ICD" and "Date"}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{ICD-10 using date}
}
\value{
Two new \code{data.table}s. 1) \code{ICD}: Uniform format diagnostic codes. 2) \code{Error}: Potential error codes.
}
\description{
Convert the ICD diagnostic codes to a uniform format
}
\details{
`IcdDxShortToDecimal` used for grouping diagnostic code to PheWAS classification (IcdDxToPheWAS). `IcdDxDecimalToShort` used for grouping to the others classification (IcdDxToCCS, IcdDxToCCSLvl, and IcdDxToComorbid). The transformative function is not only convert the ICD to uniform format code but also check potential coding error of the ICD codes’ format or version.
}
\examples{
# sample file for example

head(sampleDxFile)

# convert the diagnostic codes to the decimal format

IcdDxShortToDecimal(sampleDxFile,ICD,Date,"2015/10/01")

# convert the diagnostic codes to the short format

IcdDxDecimalToShort(sampleDxFile,ICD,Date, "2015/10/01")
}
