% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/icdDxDecimalToShort.R,
%   R/icdDxShortToDecimal.R
\name{dxUniform}
\alias{dxUniform}
\alias{icdDxDecimalToShort}
\alias{icdDxShortToDecimal}
\title{Code format transformation}
\usage{
icdDxDecimalToShort(
  dxDataFile,
  icdColName,
  dateColName,
  icdVerColName = NULL,
  icd10usingDate = NULL
)

icdDxShortToDecimal(
  dxDataFile,
  icdColName,
  dateColName,
  icdVerColName = NULL,
  icd10usingDate = NULL
)
}
\arguments{
\item{dxDataFile}{A data frame object of clinical diagnostic data with at least 2 columns: ICD, and Date. As for date column, the data format should be YYYY/MM/DD or YYYY-MM-DD.}

\item{icdColName}{Column name of ICD column in dxDataFile. Data type of this argumant should be string without quotation marks.}

\item{dateColName}{Column name of date column in dxDataFile, and the type of date column should be a date format in R or a string format with date information in YYYY/MM/DD or YYYY-MM-DD. Data type of this argumant should be string without quotation marks.}

\item{icdVerColName}{(Optional) Column name if there is a columns to record ICD-9/10 version used in dxDataFile. In this column, data format should be numeric 9L or 10L to indicate which ICD version is used for each cell. See examples below to get more information.}

\item{icd10usingDate}{The date that ICD-10 was started to be used in dxDataFile dataset. The data format should be YYYY/MM/DD or YYYY-MM-DD. Necessary if icdVerColName is null.}
}
\value{
Two new \code{data.table}s. 1) \code{ICD}: Uniform format diagnostic codes with column name "ICD". 2) \code{Error}: Potential error codes with 5 columns: ICD, count, IcdVersionInFile, WrongType and Suggestion.
}
\description{
These two functions can convert the ICD diagnostic codes to a uniform format.
}
\details{
\code{icdDxShortToDecimal} can be used for grouping diagnostic code to PheWAS classification (\code{\link{dxPheWAS}}: icdDxToPheWAS). \code{icdDxDecimalToShort} can be used for grouping to the other classification methods (\code{\link{dxCCS}}: icdDxToCCS & icdDxToCCSLvl; \code{\link{dxComorbid}}: icdDxToComorbid). These transformation functions do not only convert the ICD codes to uniform format but also check potential coding error of the ICD format and version.
}
\examples{
# sample file for example

head(sampleDxFile)

# convert the diagnostic codes to the decimal format

icdDxShortToDecimal(sampleDxFile,ICD,Date, icd10usingDate = "2015/10/01")

# convert the diagnostic codes to the short format

icdDxDecimalToShort(sampleDxFile,ICD,Date, icd10usingDate = "2015/10/01")
}
