% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConditionEra.R
\name{getConditionEra}
\alias{getConditionEra}
\title{Get the condition era}
\usage{
getConditionEra(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, gapDate = 30, icdorCCS = CCS,
  isCCSDescription = FALSE)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{Icd 10 using date}

\item{gapDate}{Length ofcondition era, default is 30 days}

\item{icdorCCS}{Stratified by icd or ccs, default is CCS}

\item{isCCSDescription}{Clinical Classifications Software (CCS) single level categories (False) and description (True) for ICD-9 or ICD-10, default is False.}
}
\description{
A Condition Era is defined as a span of time when the member is assumed to have a given condition.
Condition Eras are periods of Condition Occurrence.
Combining individual Condition Occurrences into a single Condition Era based on ICD code in clinical diagnostic data.
Condition Eras are built with a Persistence Window,deafault is 30 days, meaning, if no occurence of the same member id happens within 30 days of any one occurrence, it will be considered the end date of the last condition occurrence.
}
\details{
return DxDataFile with new column, condition era.
}
\examples{
DxDataFile <- data.frame(ID=c("A","A","A"),
                         ICD=c("6929","V433","I350"),
                         Date=as.Date(c("2013-03-31","2013-01-29","2016-03-10")),
                         stringsAsFactors = FALSE)
getConditionEra(DxDataFile, ID, ICD, Date, "2016-01-01", 30, ccs, FALSE)
getConditionEra(DxDataFile, ID, ICD, Date, "2016-01-01", 30, ICD, FALSE)

}
