% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConditionEra.R
\name{getConditionEra}
\alias{getConditionEra}
\title{Get the condition era}
\usage{
getConditionEra(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, groupDataType = ccs, CustomGroupingTable,
  isDescription = TRUE, gapDate = 30, selectedCaseFile = NULL)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{Icd 10 using date}

\item{groupDataType}{Four Stratified methods can be chosen: CCS (\code{'ccs'}), CCS levels (\code{'ccslvl1'}, \code{'ccslvl2'}, \code{'ccslvl3'}, \code{'ccslvl4'}), PheWAS (\code{'PheWAS'}), comorbidities (\code{'ahrq'},\code{'charlson'}, \code{'elix'}), grepICD or customICD (\code{'customGrepIcdGroup'}, \code{'customIcdGroup'}). Change it to any of the other possible variables, default it is set to \code{"ccs"}.}

\item{CustomGroupingTable}{Table is for groupDataType}

\item{isDescription}{CCS/PheWAS categories or description for ICD-CM codes, default is \code{'TRUE'}.}

\item{gapDate}{Length of condition era,By default it is set to 30 days \code{"30"}.}

\item{selectedCaseFile}{Table for selectedCases. Default is \code{'NULL'}}
}
\description{
A Condition Era is defined as a span of time when the member is assumed to have a given condition.
Condition Eras are periods of Condition Occurrence.
Combining individual Condition Occurrences into a single Condition Era based on ICD code in clinical diagnostic data.
Condition Eras are built with a Persistence Window,deafault is 30 days, meaning, if no occurence of the same member id happens within 30 days of any one occurrence, it will be considered the end date of the last condition occurrence.
}
\details{
return DxDataFile with new column, condition era.
}
\examples{
head(sampleDxFile)
selectedCaseFile <- selectCases(sampleDxFile, ID, ICD, Date,
                                icd10usingDate = "2015/10/01",
                                groupDataType = ccslvl3,
                                caseCondition = "Shock",
                                caseCount = 2)
getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01",
                groupDataType = CCSlvl3,
                selectedCaseFile = selectedCaseFile)

grepTable <- data.frame(group = "Cardiac dysrhythmias",
                        grepIcd = "^427|^I48",
                        stringsAsFactors = FALSE)
getConditionEra(sampleDxFile, ID, ICD, Date, "2015-10-01",
                groupDataType = customGrepIcdGroup,
                CustomGroupingTable = grepTable)

}
