% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedDataLongToWide.R
\name{groupedDataLongToWide}
\alias{groupedDataLongToWide}
\title{grouping comorbid method comorbidities (AHRQ, Charlson and Elixhauser Comorbidity) infers whether to use ICD-9 or ICD-10 codes}
\usage{
groupedDataLongToWide(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, groupedICDMethod, isDescription = T,
  CustomGroupingTable, numericOrBinary = N)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{Icd 10 using date}

\item{groupedICDMethod}{Four grouping method can be chosen: CCS (or CCS levels), phecode (only for icd-9), comorbidities, grepICD or customICD , type `ccs`,`ccslvl1`,`ccslvl2`,`ccslvl3`,`ccslvl4`, `phecode`, `ahrq`, `charlson`, `elix` `grepICD`, or `customGroup`}

\item{isDescription}{CCS/Phecode categories or description for ICD-CM codes, default is True}

\item{CustomGroupingTable}{Table is for groupedICDMethod:`grepICD` and `customGroup`}

\item{numericOrBinary}{Members have same diagnostic categories, type `N` or `B`, default is `B` (Binary)}
}
\description{
Get comorbidities using the comorbidity methods based on ICD code in clinical diagnostic data.
}
\details{
return comorbidity meseaures based on ICD diagnosis codes
}
\examples{
groupingTable <- data.frame(group = rep("Cardiac dysrhythmias",6),
                            ICD = c("427.1","427.2","427.31","427.61","427.81","427.89"),
                            stringsAsFactors = FALSE)
grepTable <- data.frame(group = c("Cardiac dysrhythmias"),
                        grepIcd = c("^427|^I48"),
                        stringsAsFactors = FALSE)
groupedDataLongToWide(testDxFile, ID, ICD, Date, "2015-10-01", ccs)
groupedDataLongToWide(testDxFile, ID, ICD, Date, "2015-10-01", ccslvl2)
groupedDataLongToWide(testDxFile, ID, ICD, Date, "2015-10-01", phecode)
groupedDataLongToWide(testDxFile, ID, ICD, Date, "2015-10-01", ahrq)
groupedDataLongToWide(testDxFile, ID, ICD, Date, "2015-10-01", charlson)
groupedDataLongToWide(testDxFile, ID, ICD, Date, "2015-10-01", elix)
groupedDataLongToWide(testDxFile, ID, ICD, Date, "2015-10-01", grepicd,
                      CustomGroupingTable = grepTable)
groupedDataLongToWide(testDxFile, ID, ICD, Date, "2015-10-01", customgroup,
                      CustomGroupingTable = groupingTable)

}
