% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedDataLongToWide.R
\name{groupedDataLongToWide}
\alias{groupedDataLongToWide}
\title{Grouped data format conversion}
\usage{
groupedDataLongToWide(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, groupDataType = ccs, CustomGroupingTable,
  isDescription = TRUE, numericOrBinary = B, selectedCaseFile = NULL)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{Icd 10 using date}

\item{groupDataType}{Four Stratified methods can be chosen: CCS (\code{'ccs'}), CCS levels (\code{'ccslvl1'}, \code{'ccslvl2'}, \code{'ccslvl3'}, \code{'ccslvl4'}), phecode (\code{'phecode'}), comorbidities (\code{'ahrq'},\code{'charlson'}, \code{'elix'}), grepICD or customICD (\code{'customGrepIcdGroup'}, \code{'customIcdGroup'}). Change it to any of the other possible variables, default it is set to \code{"ccs"}.}

\item{CustomGroupingTable}{Table is for groupDataType}

\item{isDescription}{CCS/Phecode categories or description for ICD-CM codes, default is \code{'TRUE'}.}

\item{numericOrBinary}{Members have same diagnostic categories, type `N` or `B`, default is Binary \code{'B'}.}

\item{selectedCaseFile}{Table for selectedCases. Default is \code{'NULL'}}
}
\description{
convert long format to wide format
}
\details{
return numeric or binary wide format
}
\examples{
groupingTable <- data.frame(group = rep("Cardiac dysrhythmias",6),
                            ICD = c("427.1","427.2","427.31","427.61","427.81","427.89"),
                            stringsAsFactors = FALSE)
groupedDataLongToWide(sampleDxFile, ID, ICD, Date, "2015-10-01",
                      groupDataType = customIcdGroup,
                      CustomGroupingTable = groupingTable)
selectedCaseFile <- selectCases(sampleDxFile, ID, ICD, Date,
                                icd10usingDate = "2015/10/01",
                                groupDataType = ccslvl2,
                                caseCondition = "Diseases of the heart",
                                caseCount = 2)
groupedDataLongToWide(sampleDxFile, ID, ICD, Date,
                      "2015-10-01", ccslvl2,
                      numericOrBinary = N,
                      selectedCaseFile = selectedCaseFile)

}
