% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupIcdToCCSLvl.R
\name{groupIcdBasedOnCCSLvl}
\alias{groupIcdBasedOnCCSLvl}
\title{Get the Multi-level Clinical Classifications Software (CCS) categories and description for ICD-9 and ICD-10 codes on diagnoses.}
\usage{
groupIcdBasedOnCCSLvl(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, CCSLevel = 1, CCSLvlLabel = TRUE)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{icd 10 using date}

\item{CCSLevel}{Clinical Classifications Software (CCS) multiple level}

\item{CCSLvlLabel}{Clinical Classifications Software (CCS) multiple level categories/ description for icd9/ 10, default is True}
}
\description{
Multi-leve Clinical Classifications Software (CCS) for ICD-9 and ICD-10 diagnosis codes in clinical diagnostic data is a diagnosis categorization scheme. Four levels exist in the multi-level diagnosis CCS for ICD-9-CM codes, and two levels exist in the multi-level diagnosis CCS for ICD-10-CM codes
}
\details{
return Multi-leve Clinical Classifications Software (CCS) categories or description based on ICD-9 and ICD-10 codes
}
\examples{
DxDataFile <- data.frame(ID=c("A","A","A"),
                         ICD=c("6929","V433","I350"),
                         Date=as.Date(c("2013-03-31","2013-01-29","2016-03-10")),
                         stringsAsFactors = FALSE)
groupIcdBasedOnCCSLvl (DxDataFile, ID, ICD, Date, "2016-01-01", 2, TRUE)

}
