% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMethodSelect.R
\name{groupMethodSelect}
\alias{groupMethodSelect}
\title{Select code classification method}
\usage{
groupMethodSelect(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, groupMethod, CustomGroupingTable, isDescription)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{Icd 10 using date}

\item{groupMethod}{Four Stratified methods can be chosen: CCS (\code{'ccs'}), CCS levels (\code{'ccslvl1'}, \code{'ccslvl2'}, \code{'ccslvl3'}, \code{'ccslvl4'}), phecode (\code{'phecode'}), comorbidities (\code{'ahrq'},\code{'charlson'}, \code{'elix'}), grepICD or customICD (\code{'customGrepIcdGroup'}, \code{'customIcdGroup'}). Change it to any of the other possible variables.}

\item{CustomGroupingTable}{Table is for groupDataType}

\item{isDescription}{CCS/Phecode categories or description for ICD-CM codes, default is \code{'TRUE'}.}
}
\description{
return grouped data
}
