% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{icd10_charlson}
\alias{icd10_charlson}
\title{Charlson Comorbidity for ICD-10}
\format{A data table with 2804 rows and 3 variables}
\usage{
data(icd10_charlson)
}
\description{
A datatable for ICD-10 Charlson comorbidity mapping
}
\details{
\itemize{
  \item ICD. ICD-10
  \item Comorbidity. comorbidity abbreviation measures for ICD-10 (MI, CHF, PVD, CEVD, DEM, COPD, Rheum, PUD, MILDLD, DIAB_NC, DIAB_C, PARA, RD, CANCER, MSLD, METS, HIV))
  \item Description. comorbidity measures
}
}
\seealso{
\code{\link{icd9_ahrq}} \code{\link{icd10_ahrq}} \code{\link{icd9_charlson}} \code{\link{icd9_elix}} \code{\link{icd10_elix}}
}
\keyword{datatable}
