% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{icd10_elix}
\alias{icd10_elix}
\title{Elixhauser Comorbidity for ICD-10}
\format{A data frame with 399 rows and 3 variables}
\usage{
data(icd10_elix)
}
\description{
A dataframe for ICD-10 Elixhauser comorbidity mapping
}
\details{
\itemize{
  \item ICD. ICD-10
  \item Comorbidity. comorbidity measures for ICD-10 (CHF, Arrhythmia, Valvular, PHTN, PVD, HTN, HTNcx, Paralysis, NeuroOther, Pulmonary, DM, DMcx, Hypothyroid, Renal, Liver, PUD, HIV, Lymphoma, Mets, Tumor, Rheumatic, Coagulopathy, Obesity, WeightLoss, FluidsLytes, BloodLoss, Anemia, Alcohol, Drugs, Psychoses, Depression)
  \item Value. if ICD code has comorbidity measures, value is 1
}
}
\keyword{dataframe}
