% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10GenerateDecimalFormat.R
\name{icd10GenerateDecimalFormat}
\alias{icd10GenerateDecimalFormat}
\title{get ICD-10-CM Codes Decimal Forms}
\usage{
icd10GenerateDecimalFormat(icd10)
}
\arguments{
\item{icd10}{ICD-10-CM codes}
}
\description{
Convert codes between short and decimal forms
}
