% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{icd9_ahrq}
\alias{icd9_ahrq}
\title{AHRQ Comorbidity for ICD-9}
\format{A data table with 1487 rows and 3 variables}
\usage{
data(icd9_ahrq)
}
\description{
A datatable for ICD-9 AHRQ comorbidity mapping
}
\details{
\itemize{
  \item ICD. ICD-9
  \item Comorbidity. comorbidity abbreviation measures for ICD-9 (CHF, Valvular, PHTN, PVD, HTN, HTNcx, Paralysis, NeuroOther, Pulmonary, DM, DMcx, Hypothyroid, Renal, Liver, PUD, HIV, Lymphoma, Mets, Tumor, Rheumatic, Coagulopathy, Obesity, WeightLoss, FluidsLytes, BloodLoss, Anemia, Alcohol, Drugs, Psychoses, Depression)
  \item Description. comorbidity measures
}
}
\seealso{
\code{\link{icd10_ahrq}} \code{\link{icd9_charlson}} \code{\link{icd10_charlson}} \code{\link{icd9_elix}} \code{\link{icd10_elix}}
}
\keyword{datatable}
