% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{icd9_charlson}
\alias{icd9_charlson}
\title{Charlson Comorbidity for ICD-9}
\format{A data frame with 12617 rows and 3 variables}
\usage{
data(icd9_charlson)
}
\description{
A dataframe for ICD-9 Charlson comorbidity mapping
}
\details{
\itemize{
  \item ICD. ICD-9
  \item Comorbidity. comorbidity measures for ICD-9 (MI, CHF, PVD, Stroke, Dementia, Pulmonary, Rheumatic, PUD, LiverMild, DM, DMcx, Paralysis, Renal, Cancer, LiverSevere, Mets, HIV)
  \item Value. if ICD code has comorbidity measures, value is 1
}
}
\seealso{
\code{\link{icd9_ahrq}} \code{\link{icd10_ahrq}} \code{\link{icd10_charlson}} \code{\link{icd9_elix}} \code{\link{icd10_elix}}
}
\keyword{dataframe}
