% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{icd9_elix}
\alias{icd9_elix}
\title{Elixhauser Comorbidity for ICD-9}
\format{A data frame with 1251 rows and 4 variables}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp#references}
}
\usage{
data(icd9_elix)
}
\description{
A dataframe for ICD-9 Elixhauser comorbidity mapping
}
\details{
\itemize{
  \item ICD. ICD-9
  \item Comorbidity_abbr. comorbidity measures abbreviation for ICD-9
  \item Comorbidity. comorbidity measures for ICD-9 (CHF, VALVE, PULMCIRC, PERIVASC, HTN_C (using HTN, HTNCX), PARA, NEURO, CHRNLUNG, DM, DMcx, HYPOTHY, RENLFAIL, LIVER, ULCER, AIDS, Lymph, METS, TUMOR, ARTH, Tumor, COAG, OBESE, WGHTLOSS, LYTES, BLDLOSS, ANEMDEF, ALCOHOL, DRUG, PSYCH, DEPESS)
  \item Value. if ICD code has comorbidity measures, value is 1
}
}
\seealso{
\code{\link{icd9_ahrq}} \code{\link{icd10_ahrq}} \code{\link{icd9_charlson}} \code{\link{icd10_charlson}} \code{\link{icd10_elix}}
}
\keyword{dataframe}
