% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICD9PrGenerateDecimalFormat.R
\name{ICD9PrGenerateDecimalFormat}
\alias{ICD9PrGenerateDecimalFormat}
\title{get ICD-9-PCS Codes Decimal Forms}
\source{
\url{https://www.findacode.com/search/search.php}
}
\usage{
ICD9PrGenerateDecimalFormat(icd9)
}
\arguments{
\item{icd9}{ICD-9-PCS codes}
}
\description{
Convert codes between short and decimal forms
Format ICD-9-PCS Principal and Other Diagnosis Codes can be found at url link.
}
