% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{ICD9PrwithTwoFormat}
\alias{ICD9PrwithTwoFormat}
\title{Two format of ICD-9-PCS: Short and Decimal}
\format{A data frame with 3948 rows and 2 variables}
\source{
\url{https://www.findacode.com/search/search.php}
}
\usage{
data(ICD9PrwithTwoFormat)
}
\description{
A dataframe for ICD-9 format conversion
}
\details{
\itemize{
  \item Short ICD-9 Short format
  \item Decimal ICD-9 Decimal format
}
}
\keyword{dataframe}
