% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxDecimalToShort.R
\name{IcdDxDecimalToShort}
\alias{IcdDxDecimalToShort}
\title{Convert ICD Codes From Decimal To Short Forms}
\source{
\url{https://www.findacode.com/search/search.php}

\url{https://www.cms.gov/Medicare/Quality-Initiatives-Patient-Assessment-Instruments/HospitalQualityInits/Downloads/HospitalAppendix_F.pdf}

\url{https://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}

\url{https://www.cms.gov/Medicare/Coding/ICD10/2019-ICD-10-CM.html}
}
\usage{
IcdDxDecimalToShort(DxDataFile, icdColName, dateColName, icd10usingDate)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 2 columns: "ICD", "Date"}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{icd 10 using date}
}
\description{
Convert codes between short and decimal forms
}
\examples{
IcdDxDecimalToShort(sampleDxFile,ICD,Date,"2015/10/01")
}
