% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxShortToDecimal.R
\name{IcdDxShortToDecimal}
\alias{IcdDxShortToDecimal}
\title{Convert ICD Codes From Short To Decimal Forms}
\usage{
IcdDxShortToDecimal(icdList)
}
\arguments{
\item{icdList}{ICD codes}
}
\description{
Convert codes between short and decimal forms
}
