% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxShortToDecimal.R
\name{IcdDxShortToDecimal}
\alias{IcdDxShortToDecimal}
\title{Convert ICD Codes From Short To Decimal Forms}
\source{
\url{https://www.findacode.com/search/search.php}

\url{https://www.cms.gov/Medicare/Quality-Initiatives-Patient-Assessment-Instruments/HospitalQualityInits/Downloads/HospitalAppendix_F.pdf}

\url{https://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}

\url{https://www.cms.gov/Medicare/Coding/ICD10/2019-ICD-10-CM.html}
}
\usage{
IcdDxShortToDecimal(icdList)
}
\arguments{
\item{icdList}{ICD codes}
}
\description{
Convert codes between short and decimal forms
}
