% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdDxToComorbid.R
\name{IcdDxToComorbid}
\alias{IcdDxToComorbid}
\title{Grouping comorbid method comorbidities (AHRQ, Charlson and Elixhauser Comorbidity) infers whether to use ICD-9 or ICD-10 codes}
\source{
AHRQ

Charlson

Elixhauser

ICD-9-CM Elixhauser (2012-2015)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp#references}

ICD-10-CM Elixhauser (2019)

\url{https://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity_icd10.jsp}
}
\usage{
IcdDxToComorbid(DxDataFile, idColName, icdColName, dateColName,
  icd10usingDate, comorbidMethod)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{ICD 10 using date}

\item{comorbidMethod}{Three comorbidity method: AHRQ, Charlson and Elixhauser Comorbidity. Change
it to any of the other possible variables (\code{'ahrq'},\code{'charlson'}, \code{'elix'}).}
}
\description{
Get comorbidities using the comorbidity methods based on ICD code in clinical diagnostic data.
}
\details{
return comorbidity meseaures based on ICD diagnosis codes
}
\examples{
head(sampleDxFile)
IcdDxToComorbid(sampleDxFile, ID, ICD, Date, "2015-10-01", charlson)
IcdDxToComorbid(sampleDxFile, ID, ICD, Date, "2015-10-01", elix)
IcdDxToComorbid(sampleDxFile, ID, ICD, Date, "2015-10-01", ahrq)

}
