% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrDecimaltoShort.R
\name{IcdPrDecimalToShort}
\alias{IcdPrDecimalToShort}
\title{Convert ICD-9-PCS Codes From Decimal To Short Forms}
\source{
\url{https://www.findacode.com/search/search.php}

\url{https://www.cms.gov/Medicare/Quality-Initiatives-Patient-Assessment-Instruments/HospitalQualityInits/Downloads/HospitalAppendix_F.pdf}

\url{https://www.cms.gov/Medicare/Coding/ICD10/2019-ICD-10-PCS.html}
}
\usage{
IcdPrDecimalToShort(icdList)
}
\arguments{
\item{icdList}{only ICD-9-PCS codes have decimal and short forms}
}
\description{
Return codes with short forms
}
