% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrDecimaltoShort.R
\name{IcdPrDecimaltoShort}
\alias{IcdPrDecimaltoShort}
\title{Convert ICD-9-PCS Codes From Decimal To Short Forms}
\source{
\url{https://www.findacode.com/search/search.php}

\url{https://www.cms.gov/Medicare/Quality-Initiatives-Patient-Assessment-Instruments/HospitalQualityInits/Downloads/HospitalAppendix_F.pdf}
}
\usage{
IcdPrDecimaltoShort(icdList)
}
\arguments{
\item{icdList}{only ICD-9-PCS codes have decimal and short forms}
}
\description{
Return codes with short forms
}
