% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrDecimaltoShort.R
\name{IcdPrDecimalToShort}
\alias{IcdPrDecimalToShort}
\title{Convert ICD-9-PCS Codes From Decimal To Short Forms}
\source{
\url{https://www.findacode.com/search/search.php}

\url{https://www.cms.gov/Medicare/Quality-Initiatives-Patient-Assessment-Instruments/HospitalQualityInits/Downloads/HospitalAppendix_F.pdf}

\url{https://www.cms.gov/Medicare/Coding/ICD10/2019-ICD-10-PCS.html}
}
\usage{
IcdPrDecimalToShort(PrDataFile, icdColName, dateColName, icd10usingDate)
}
\arguments{
\item{PrDataFile}{A file of clinical diagnostic data with at least 2 columns: "ICD", "Date"}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{icd 10 using date}
}
\description{
Return codes with short forms
}
\examples{
head(samplePrFile)
samplePrFile$Short <- IcdPrDecimalToShort(samplePrFile,ICD,Date,"2015/10/01")$ICD

}
