% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrShortToDecimal.R
\name{IcdPrShortToDecimal}
\alias{IcdPrShortToDecimal}
\title{Convert ICD-9-PCS Codes From Short To Decimal Forms}
\source{
\url{https://www.findacode.com/search/search.php}

\url{https://www.cms.gov/Medicare/Quality-Initiatives-Patient-Assessment-Instruments/HospitalQualityInits/Downloads/HospitalAppendix_F.pdf}
}
\usage{
IcdPrShortToDecimal(icdList)
}
\arguments{
\item{icdList}{only ICD-9-PCS codes have decimal and short forms}
}
\description{
Return codes with Decimal forms
}
