% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrShortToDecimal.R
\name{IcdPrShortToDecimal}
\alias{IcdPrShortToDecimal}
\title{Convert ICD-9-PCS Codes From Short To Decimal Forms}
\source{
\url{https://www.findacode.com/search/search.php}

\url{https://www.cms.gov/Medicare/Quality-Initiatives-Patient-Assessment-Instruments/HospitalQualityInits/Downloads/HospitalAppendix_F.pdf}

\url{https://www.cms.gov/Medicare/Coding/ICD10/2019-ICD-10-PCS.html}
}
\usage{
IcdPrShortToDecimal(PrDataFile, icdColName, dateColName, icd10usingDate)
}
\arguments{
\item{PrDataFile}{A file of clinical diagnostic data with at least 2 columns: "ICD", "Date"}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{icd 10 using date}
}
\description{
Return codes with Decimal forms
}
\examples{
head(samplePrFile)
IcdPrShortToDecimal(samplePrFile,ICD,Date,"2015/10/01")

}
