% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrToProcedureClass.R
\name{IcdPrToProcedureClass}
\alias{IcdPrToProcedureClass}
\title{Get the Procedure Class for ICD-9 and ICD-10 codes on procedures.}
\usage{
IcdPrToProcedureClass(PrDataFile, idColName, icdColName, dateColName,
  icd10usingDate, isProcedureClassName = TRUE)
}
\arguments{
\item{PrDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of PrDataFile}

\item{icdColName}{A column for ICD of PrDataFile}

\item{dateColName}{A column for Date of PrDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{isProcedureClassName}{Procedure Class category/name for ICD-9 or ICD-10, default is True}
}
\description{
return Procedure Class categories based on ICD-9 and ICD-10 codes
}
\examples{
PrDataFile <- data.frame(ID=c("A","A","A","B"),
                         ICD=c("0101","8838","00870ZZ","00920ZZ"),
                         Date=as.Date(c("2013-03-31","2013-01-29","2016-03-10","2016-03-10")),
                         stringsAsFactors = FALSE)
IcdPrToProcedureClass(PrDataFile, ID, ICD, Date, "2016-01-01", TRUE)

}
