% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/icdPrToProcedureClass.R
\name{PC}
\alias{PC}
\alias{icdPrToProcedureClass}
\title{Code classification for procedure code}
\usage{
icdPrToProcedureClass(
  prDataFile,
  idColName,
  icdColName,
  dateColName,
  icdVerColName = NULL,
  icd10usingDate = NULL,
  isDescription = TRUE
)
}
\arguments{
\item{prDataFile}{A data frame object of clinical procedure data with at least 3 columns: ID, ICD, and Date. As for date column, the data format should be YYYY/MM/DD or YYYY-MM-DD.}

\item{idColName}{Column name of ID column in prDataFile. Data type of this argumant should be string without quotation marks.}

\item{icdColName}{Column name of ICD column in prDataFile. Data type of this argumant should be string without quotation marks.}

\item{dateColName}{Column name of date column in prDataFile (with date in YYYY/MM/DD or YYYY-MM-DD format). Data type of this argumant should be string without quotation marks.}

\item{icd10usingDate}{ICD-10 using date}

\item{isDescription}{Category or description of standard classification methods for ICD procedure codes. By default, it set to \code{True} (standard description).}
}
\value{
Two new \code{data.table}s. 1) \code{groupedDT}: Based on \code{prDataFile} with two new columns for uniform format procedure codes and classified standard categories. 2) \code{Error}: Potential error codes from \code{\link{PrUniform}}.
}
\description{
The Procedure Class classification for ICD-9 and ICD-10 codes is a procedure categorization scheme that can employ in many types of projects analyzing data
}
\details{
Collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than are individual ICD procedure codes.
}
\examples{
# sample file for example

head(samplePrFile)

# Group procedure codes into procedure class classification

icdPrToProcedureClass(samplePrFile, ID, ICD, Date, icdVerColName = NULL, "2015-10-01", TRUE)

}
\seealso{
see other code classification: \code{\link{PrCCS}}
}
