% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_errorICD.R
\name{plot_errorICD}
\alias{plot_errorICD}
\title{Pareto Plot of error ICD Codes}
\usage{
plot_errorICD(errorFile, ICDVersion = all, wrongICDType = all,
  groupICD = FALSE, Others = TRUE, TopN = 10)
}
\arguments{
\item{errorFile}{error file from ICD uniform function (`IcdDxDecimalToShort` or `IcdDxShortToDecimal`)}

\item{ICDVersion}{ICD version}

\item{wrongICDType}{Wrong ICD type}

\item{groupICD}{Only ICD-9 codes can be grouped, because ICD 10 already has unique alphanumeric codes to identify known diseases. Default is FALSE}

\item{Others}{Default is TRUE}

\item{TopN}{Default is Top "10"}
}
\description{
Pareto Plot of error ICD Codes
}
\examples{
head(sampleDxFile)
error <- IcdDxDecimalToShort(sampleDxFile, ICD, Date, "2015/10/01")
plot_errorICD(errorFile = error$Error,
              ICDVersion = 9,
              wrongICDType = all,
              groupICD = TRUE,
              Others = TRUE,
              TopN = 3)

plot_errorICD(errorFile = error$Error,
              ICDVersion = all,
              wrongICDType = all,
              groupICD = FALSE,
              Others = TRUE)

}
