% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_groupedData.R
\name{plot_groupedData}
\alias{plot_groupedData}
\title{Plot of diagnostic category data}
\usage{
plot_groupedData(groupedDataWide, TopN = 10, limitPrevalance = 0.01,
  pvalue = 0.05)
}
\arguments{
\item{groupedDataWide}{groupedData file from functions of code classification (four strategies)}

\item{TopN}{Default is Top "10"}

\item{limitPrevalance}{the minimum  of prevalance Default is "0.01", in other words,the limit at same diagnostic category must have 1 percent patient in total patient.}

\item{pvalue}{p value of chisq.test}
}
\description{
Plot of diagnostic category data
}
\examples{
head(sampleDxFile)
groupedDataWide <- groupedDataLongToWide(sampleDxFile, ID, ICD, Date,
                                         icd10usingDate = "2015-10-01",
                                         groupDataType = elix)
head(groupedDataWide)
plot_groupedData(groupedDataWide = groupedDataWide,
                 TopN = 10,
                 limitPrevalance = 0.01,
                 pvalue = 0.001)

}
