% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_emr.R
\docType{data}
\name{Pr_dataset_standardization}
\alias{Pr_dataset_standardization}
\alias{ICD9PrwithTwoFormat}
\alias{prICD10}
\title{Datasets for code standardization}
\format{
These datatable with ICD9 and ICD10 procedure codes.

An object of class \code{data.table} (inherits from \code{data.frame}) with 78880 rows and 2 columns.
}
\usage{
ICD9PrwithTwoFormat

prICD10
}
\description{
These datasets are used for procedure code standardization
}
\details{
These datasets are used for \code{IcdPrShortToDecimal}, and \code{IcdPrDecimalToShort}
}
\keyword{datasets}
