% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IcdPrToCCS.R, R/IcdPrToCCSLvl.R, R/fun_emr.R
\name{IcdPrToCCS}
\alias{IcdPrToCCS}
\alias{IcdPrToCCSLvl}
\alias{PrCCS}
\title{Code classification for CCS}
\usage{
IcdPrToCCS(PrDataFile, idColName, icdColName, dateColName, icd10usingDate,
  isDescription = TRUE)

IcdPrToCCSLvl(PrDataFile, idColName, icdColName, dateColName,
  icd10usingDate, CCSLevel = 1, isDescription = TRUE)
}
\arguments{
\item{PrDataFile}{A file of clinical procedure data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A MemberID column of PrDataFile}

\item{icdColName}{A ICD column of PrDataFile}

\item{dateColName}{A Date column of PrDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{isDescription}{Category or description of standard classification methods for ICD procedure codes. By default, it set to \code{True} (standard description).}

\item{CCSLevel}{By default, it set to \code{1}. CCS multiple level:1~3, CCS for ICD-10-CM only has 1~2 multiple levels.}
}
\value{
Two new \code{data.table}s. 1) \code{groupedDT}: Based on \code{PrDataFile} with two new columns for uniform format procedure codes and classified standard categories. 2) \code{Error}: Potential error codes from \code{\link{PrUniform}}.
}
\description{
The CCS classification for ICD-9 and ICD-10 codes is a procedure categorization scheme that can employ in many types of projects analyzing data on procedures
}
\details{
These CCS functions collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than are individual ICD procedure codes.
}
\examples{
# sample file for example

head(samplePrFile)

# Group procedure codes into single level of CCS classification

IcdPrToCCS(samplePrFile, ID, ICD, Date, "2015-10-01", TRUE)

# Group procedure codes into multiple levels of CCS classification

IcdPrToCCSLvl(samplePrFile, ID, ICD, Date, "2015-10-01", 2, TRUE)

}
\seealso{
see other code classification: \code{\link{PC}}
}
