% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/icdPrToCCS.R, R/icdPrToCCSLvl.R
\name{prCCS}
\alias{prCCS}
\alias{icdPrToCCS}
\alias{icdPrToCCSLvl}
\title{Code classification for CCS}
\usage{
icdPrToCCS(prDataFile, idColName, icdColName, dateColName,
  icdVerColName = NULL, icd10usingDate = NULL, isDescription = TRUE)

icdPrToCCSLvl(prDataFile, idColName, icdColName, dateColName,
  icdVerColName = NULL, icd10usingDate = NULL, CCSLevel = 1,
  isDescription = TRUE)
}
\arguments{
\item{idColName}{A MemberID column of PrDataFile}

\item{icdColName}{A ICD column of PrDataFile}

\item{dateColName}{A Date column of PrDataFile}

\item{icd10usingDate}{ICD-10 using date}

\item{isDescription}{Category or description of standard classification methods for ICD procedure codes. By default, it set to \code{True} (standard description).}

\item{CCSLevel}{By default, it set to \code{1}. CCS multiple level:1~3, CCS for ICD-10-CM only has 1~2 multiple levels.}
}
\value{
Two new \code{data.table}s. 1) \code{groupedDT}: Based on \code{PrDataFile} with two new columns for uniform format procedure codes and classified standard categories. 2) \code{Error}: Potential error codes from \code{\link{prUniform}}.
}
\description{
The CCS classification for ICD-9 and ICD-10 codes is a procedure categorization scheme that can employ in many types of projects analyzing data on procedures
}
\details{
These CCS functions collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than are individual ICD procedure codes.
}
\examples{
# sample file for example

head(samplePrFile)

# Group procedure codes into single level of CCS classification

icdPrToCCS(samplePrFile, ID, ICD, Date, icd10usingDate = "2015-10-01", isDescription = TRUE)

# Group procedure codes into multiple levels of CCS classification

icdPrToCCSLvl(samplePrFile, ID, ICD, Date, icd10usingDate = "2015-10-01", CCSLevel = 2, isDescription = TRUE)

}
\seealso{
see other code classification: \code{\link{PC}}
}
