% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/icdPrToCCS.R, R/icdPrToCCSLvl.R
\name{prCCS}
\alias{prCCS}
\alias{icdPrToCCS}
\alias{icdPrToCCSLvl}
\title{Code classification for CCS}
\usage{
icdPrToCCS(
  prDataFile,
  idColName,
  icdColName,
  dateColName,
  icdVerColName = NULL,
  icd10usingDate = NULL,
  isDescription = TRUE
)

icdPrToCCSLvl(
  prDataFile,
  idColName,
  icdColName,
  dateColName,
  icdVerColName = NULL,
  icd10usingDate = NULL,
  CCSLevel = 1,
  isDescription = TRUE
)
}
\arguments{
\item{prDataFile}{A data frame object of clinical procedure data with at least 3 columns: ID, ICD, and Date. As for date column, the data format should be YYYY/MM/DD or YYYY-MM-DD.}

\item{idColName}{Column name of ID column in prDataFile. Data type of this argumant should be string without quotation marks.}

\item{icdColName}{Column name of ICD column in prDataFile. Data type of this argumant should be string without quotation marks.}

\item{dateColName}{Column name of date column in prDataFile (with date in YYYY/MM/DD or YYYY-MM-DD format). Data type of this argumant should be string without quotation marks.}

\item{icdVerColName}{(Optional) Column name of ICD-9/10 version recorded in prDataFile. Data format in this column should be numeric 9L or 10L.}

\item{icd10usingDate}{The date that ICD-10 was started to be used in prDataFile dataset. The data format should be YYYY/MM/DD or YYYY-MM-DD. Necessary if icdVerColName is null.}

\item{isDescription}{Binary. If true, category description of classification methods will be used in the group column. If false, category name will be used. By default, it is set to be \code{True} (standard category description).}

\item{CCSLevel}{Numeric. Used for multi-level CCS. By default, it is set to \code{1}. There is 3 multi-level CCS (1~3) for ICD-9, and 2 multi-level CCS (1 and 2) for ICD-10.}
}
\value{
Two new \code{data.table}s. 1) \code{groupedDT}: Based on \code{prDataFile} with two new columns for uniform format procedure codes and classified standard categories. 2) \code{Error}: Potential error codes from \code{\link{prUniform}}.
}
\description{
The CCS classification for ICD-9 and ICD-10 codes is a procedure categorization scheme that can employ in many types of projects analyzing data on procedures
}
\details{
These CCS functions collapse ICD codes into a smaller number of clinically meaningful categories that are more useful for presenting descriptive statistics than individual ICD procedure codes are.
}
\examples{
# sample file for example

head(samplePrFile)

# Group procedure codes into single level of CCS classification

icdPrToCCS(samplePrFile, ID, ICD, Date, icdVerColName = NULL, "2015-10-01", TRUE)

# Group procedure codes into multiple levels of CCS classification

icdPrToCCSLvl(samplePrFile, ID, ICD, Date, icdVerColName = NULL, "2015-10-01", 2, TRUE)

}
\seealso{
see other code classification: \code{\link{PC}} (icdPrToProcedureClass).
}
