% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/icdPrDecimalToShort.R,
%   R/icdPrShortToDecimal.R
\name{prUniform}
\alias{prUniform}
\alias{icdPrDecimalToShort}
\alias{icdPrShortToDecimal}
\title{Code format transformation}
\usage{
icdPrDecimalToShort(
  prDataFile,
  icdColName,
  dateColName,
  icdVerColName = NULL,
  icd10usingDate = NULL
)

icdPrShortToDecimal(
  prDataFile,
  icdColName,
  dateColName,
  icdVerColName = NULL,
  icd10usingDate = NULL
)
}
\arguments{
\item{prDataFile}{A file of clinical procedure data with at least 2 columns: "ICD" and "Date"}

\item{icdColName}{A ICD column of prDataFile}

\item{dateColName}{A Date column of prDataFile}

\item{icd10usingDate}{ICD-10 using date}
}
\value{
Two new \code{data.table}s. 1) \code{ICD}: Uniform format procedure codes. 2) \code{Error}: Potential error codes.
}
\description{
Converts the ICD procedure codes to a uniform format
}
\details{
These transformative function does not only convert the ICD to uniform format code but also check the potential coding error of the ICD codes’ format or version.
}
\examples{
# sample file for example

head(samplePrFile)

# convert the procedure codes to the short format

icdPrDecimalToShort(samplePrFile,ICD,Date, icdVerColName = NULL, "2015/10/01")

# convert the procedure codes to the decimal format

icdPrShortToDecimal(samplePrFile,ICD,Date, icdVerColName = NULL, "2015/10/01")
}
