% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/patientRecordDate.R
\name{recordPeriod}
\alias{recordPeriod}
\alias{patientRecordDate}
\title{Data integration for patient record period}
\usage{
patientRecordDate(DxDataFile, idColName, icdColName, dateColName)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", and "Date"}

\item{idColName}{A column for MemberID of DxDataFile.}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}
}
\value{
A new \code{data.table} based on \code{DxDataFile} with the earliest and latest admission date for each patient.
}
\description{
Function for researchers used for finding the first and last clinical event for a given patient as index date.
}
\details{
The function queries the earliest and latest admission date for each patient.
}
\examples{
# sample file for example

head(sampleDxFile)

# Earliest and latest admission date for each patient

record <- patientRecordDate(sampleDxFile, ID, ICD, Date)
head(record)
}
\seealso{
Other data integration functions: \code{\link{selectCase}}, \code{\link{DataSplit}}, \code{\link{era}}
}
