% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{sampleDxFile}
\alias{sampleDxFile}
\title{Sample File of ICD-CM codes for grouping method}
\format{A sample data table with 300 rows and 3 variables}
\usage{
data(sampleDxFile)
}
\description{
A sample file for demo with diagnostic function.
}
\details{
\itemize{
  \item ID  MemberID
  \item ICD ICD-9 and ICD-10-CM codes
  \item Date diagnosis date
}
}
\keyword{datatable}
