% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr.R
\docType{data}
\name{samplePrFile}
\alias{samplePrFile}
\title{Sample File of ICD-PCS codes for grouping method}
\format{A data table with 4010 rows and 3 variables}
\usage{
data(samplePrFile)
}
\description{
A sample file for grouping method (CCS, CCSLvl, phecode, comorbidity...) and calculating condition era
}
\details{
\itemize{
  \item ID  MemberID
  \item ICD ICD-9 and ICD-10-PCS codes
  \item Date diagnosis date
}
}
\keyword{datatable}
