% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCases.R
\name{selectCases}
\alias{selectCases}
\title{Select cases based on ICD code and the number of ICD codes}
\usage{
selectCases(grepICD, DxDataFile, idColName, icdColName, dateColName,
  ICDNumber, minimumINRofDays = 30, maximumINRofDays = 365)
}
\arguments{
\item{grepICD}{ICD selection rules with grepl expression}

\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{ICDNumber}{a threshold of number of ICD for case selection}

\item{minimumINRofDays}{Minimum interval of Days, defaults is 30 days (one month)}

\item{maximumINRofDays}{Maximum interval of Days, defaults is 365 days (one year)}
}
\description{
This can be used to select qualified cases from factIcd data
based on the ICD code searching criteria and number of ICD code
per patients in the inout factIcd dataset.
Return qualified Members' data
}
\examples{
head(sampleDxFile)
selectCases("^785", sampleDxFile, ID, ICD, Date, 2)

}
