% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCases.R
\name{selectCases}
\alias{selectCases}
\title{Select cases based on ICD code and the number of ICD codes}
\usage{
selectCases(grepICD, DxDataFile, idColName, icdColName, dateColName,
  ICDNumber, INRofDayRange = c(30, 365), selectedCaseType = "selected")
}
\arguments{
\item{grepICD}{ICD selection rules with grepl expression}

\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{ICDNumber}{a threshold of number of ICD for case selection}

\item{INRofDayRange}{Determines what is the interval of days of interest for performing the case selection. By default it is set to from 30 to 365 days.}

\item{selectedCaseType}{Aggregation  of selected cases name. By default it is set to \code{"selected"}.}
}
\description{
This can be used to select qualified cases from factIcd data
based on the ICD code searching criteria and number of ICD code
per patients in the inout factIcd dataset.
Return qualified Members' data
}
\examples{
head(sampleDxFile)
selectCases("^785", sampleDxFile, ID, ICD, Date, 2)

}
