% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCases.R
\name{selectCases}
\alias{selectCases}
\title{Select cases based on ICD code and the number of ICD codes}
\usage{
selectCases(DxDataFile, idColName, icdColName, dateColName, icd10usingDate,
  groupDataType = ICD, CustomGroupingTable, isDescription = TRUE,
  caseCondition, caseCount, INRofDayRange = c(30, 365),
  selectCaseType = "Selected")
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID","ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{icd10usingDate}{Icd 10 using date}

\item{groupDataType}{Four Stratified methods can be chosen: CCS (\code{'ccs'}), CCS levels (\code{'ccslvl1'}, \code{'ccslvl2'}, \code{'ccslvl3'}, \code{'ccslvl4'}), PheWAS (\code{'PheWAS'}), comorbidities (\code{'ahrq'},\code{'charlson'}, \code{'elix'}), grepICD or customICD (\code{'customGrepIcdGroup'}, \code{'customIcdGroup'}). Change it to any of the other possible variables, default it is set to \code{"ccs"}.}

\item{CustomGroupingTable}{Table is for groupDataType}

\item{isDescription}{CCS/PheWAS categories or description for ICD-CM codes, default is \code{'TRUE'}.}

\item{caseCondition}{ICD selection rules with grepl expression}

\item{caseCount}{a threshold of number of ICD for case selection}

\item{INRofDayRange}{Determines the interval of days of interest for performing the case selection. By default it is set from 30 to 365 days.}

\item{selectCaseType}{Aggregation  of selected cases name. By default it is set to \code{"selected"}.}
}
\description{
This can be used to select qualified cases from factIcd data
based on the ICD code searching criteria and number of ICD code
per patients in the inout factIcd dataset.
Return qualified Members' data
}
\examples{
head(sampleDxFile)
selectCases(DxDataFile = sampleDxFile,
            ID, ICD, Date,
            groupDataType = ccslvl2,
            icd10usingDate = "2015/10/01",
            caseCondition = "Diseases of the heart",
            caseCount = 2)

}
