% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDataByDate.R
\name{splitDataByDate}
\alias{splitDataByDate}
\title{split data based on index date}
\usage{
splitDataByDate(DxDataFile, idColName, icdColName, dateColName,
  indexDateFile, windowGap = 30)
}
\arguments{
\item{DxDataFile}{A file of clinical diagnostic data with at least 3 columns: "MemberID", "ICD", "Date"}

\item{idColName}{A column for MemberID of DxDataFile}

\item{icdColName}{A column for ICD of DxDataFile}

\item{dateColName}{A column for Date of DxDataFile}

\item{indexDateFile}{An exact date of diagnosis for a period of observation.}

\item{windowGap}{gap length of window. Default is set to \code{30}.}
}
\description{
split data based on index date
}
\examples{
sampleDxFile <- sampleDxFile[grepl("A0|B0|C0|D0",ID),]
head(sampleDxFile)
indexDateTable <- data.frame(ID = c("A0","B0","C0","D0"),
                             indexDate = c("2009-07-25", "2015-12-26",
                                           "2015-12-05", "2017-01-29"),
                             stringsAsFactors = FALSE)
splitDataByDate(sampleDxFile, ID, ICD, Date,
                indexDateFile = indexDateTable,
                windowGap = 30)

}
