% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_emr.R, R/splitDataByDate.R
\name{splitDataByDate}
\alias{splitDataByDate}
\title{Data integration for data split}
\usage{
splitDataByDate(
  dxDataFile,
  idColName,
  icdColName,
  dateColName,
  indexDateFile,
  gap = 30
)
}
\arguments{
\item{dxDataFile}{A data frame object of clinical diagnostic data with at least 3 columns: ID, ICD, and Date. As for date column, the data format should be YYYY/MM/DD or YYYY-MM-DD.}

\item{idColName}{Column name of ID column in dxDataFile. Data type of this argumant should be string without quotation marks.}

\item{icdColName}{Column name of ICD column in dxDataFile. Data type of this argumant should be string without quotation marks.}

\item{dateColName}{Column name of date column in dxDataFile (with date in YYYY/MM/DD or YYYY-MM-DD format). Data type of this argumant should be string without quotation marks.}

\item{indexDateFile}{A data frame contained index dates for each patient in an observed period. The column names should be 'ID' and 'indexDate'.}

\item{gap}{Gap length of the window. Default is \code{30} (data will be seperated every 30 days).}
}
\value{
A new \code{data.table} based on \code{dxDataFile} and classified by \code{indexDateFile} for each patient
}
\description{
This function splits data by the date of the clinical event and shows the data recorded before or after the clinical event, and calculates the period between the record date and index date based on a self-defined window gap.
}
\details{
In most condition, users need to extract data by a specific clinical event (e.g., first diagnosis dates of chronic diseases). Users can define a table of clinical index dates of each patient. The date can be generated by \code{\link{selectCases}} or first/last admission date by \code{\link{getEligiblePeriod}}.
}
\examples{
# sample file for example

SampleforCertainPatient <- sampleDxFile[grepl("A0|B0|C0|D0",ID),]

head(SampleforCertainPatient)

# Defined index date of patient A0,B0,C0 and D0

indexDateTable <- data.frame(ID = c("A0","B0","C0","D0"),
                             indexDate = c("2023-08-12", "2024-02-12",
                                           "2015-12-05", "2017-01-29"),
                             stringsAsFactors = FALSE)
indexDateTable
# Split data by index date for each patient

splitedData <- splitDataByDate(SampleforCertainPatient, ID, ICD, Date,
                               indexDateFile = indexDateTable,
                               gap = 30)
splitedData[15:19,]
}
\seealso{
Other data integration functions: \code{\link{selectCases}}, \code{\link{getEligiblePeriod}}, \code{\link{getConditionEra}}
}
